\docType{data}
\name{abbreviations}
\alias{abbreviations}
\title{Small Abbreviations Data Set}
\format{A data frame with 14 rows and 2 variables}
\description{
  A dataset containing abbreviations and their qdap
  friendly form.
}
\details{
  \itemize{ \item abv. Common transcript abbreviations
  \item rep. qdap representation of those abbreviations }
}
\keyword{datasets}

