\docType{data}
\name{action.verbs}
\alias{action.verbs}
\title{Action Word List}
\format{A vector with 1569 elements}
\description{
  A dataset containing a vector of action words.  This is a
  subset of the \href{http://icon.shef.ac.uk/Moby/}{Moby
  project: Moby Part-of-Speech}.
}
\details{
  \href{http://icon.shef.ac.uk/Moby/}{From Grady Ward's
  Moby project:} "This second edition is a particularly
  thorough revision of the original Moby Part-of-Speech.
  Beyond the fifteen thousand new entries, many thousand
  more entries have been scrutinized for correctness and
  modernity. This is unquestionably the largest P-O-S list
  in the world. Note that the many included phrases means
  that parsing algorithms can now tokenize in units larger
  than a single word, increasing both speed and accuracy."
}
\references{
  \url{http://icon.shef.ac.uk/Moby/mpos.html}
}
\keyword{datasets}

