% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_s.R
\name{add_s}
\alias{add_s}
\title{Make Plural (or Verb to Singular) Versions of Words}
\usage{
add_s(x, keep.original = TRUE)
}
\arguments{
\item{x}{A vector of words to make plural.}

\item{keep.original}{logical.  If \code{TRUE} the original words are kept in 
the return vector.}
}
\value{
Returns a vector of plural words.
}
\description{
Add -s, -es, or -ies to words.
}
\examples{
set.seed(10)
add_s(sample(GradyAugmented, 10))
set.seed(10)
add_s(sample(GradyAugmented, 10), FALSE)
}
\keyword{plural}
