% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Network.R
\name{+.Network}
\alias{+.Network}
\title{Add themes to a Network object.}
\usage{
\method{+}{Network}(Network.obj, x)
}
\arguments{
\item{Network.obj}{An object of class \code{Network}.}

\item{x}{A component to add to \code{Network.obj}}
}
\description{
This operator allows you to add themes to a Network object.
}

