\name{Animate}
\alias{Animate}
\title{Generic Animate Method}
\usage{
Animate(x, ...)
}
\arguments{
  \item{x}{A qdap object (list) with a animatable
  graph-plot object (e.g.,
  \code{\link[qdap]{discourse_map}}).}

  \item{\ldots}{Arguments passed to Animate method of other
  classes.}
}
\value{
Returns a plot object.
}
\description{
Animate graph-plot objects from select qdap outputs.
}
\seealso{
\code{\link[qdap]{scores}}, \code{\link[qdap]{counts}},
\code{\link[qdap]{preprocessed}},
\code{\link[qdap]{proportions}}
}

