\docType{data}
\name{BuckleySaltonSWL}
\alias{BuckleySaltonSWL}
\title{Buckley & Salton Stopword List}
\format{A character vector with 546 elements}
\description{
  A stopword list containing a character vector of
  stopwords.
}
\details{
  \href{http://www.lextek.com/manuals/onix/stopwords2.html}{From
  Onix Text Retrieval Toolkit API Reference}: "This
  stopword list was built by Gerard Salton and Chris
  Buckley for the experimental SMART information retrieval
  system at Cornell University. This stopword list is
  generally considered to be on the larger side and so when
  it is used, some implementations edit it so that it is
  better suited for a given domain and audience while
  others use this stopword list as it stands."
}
\note{
  Reduced from the original 571 words to 546.
}
\references{
  \url{http://www.lextek.com/manuals/onix/stopwords2.html}
}
\keyword{datasets}

