% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build_qdap_vignette.R
\name{build_qdap_vignette}
\alias{build_qdap_vignette}
\title{Replace Temporary Introduction to qdap Vignette}
\usage{
build_qdap_vignette(download.html = FALSE)
}
\arguments{
\item{download.html}{logical.  If \code{TRUE} the file will be downloaded
from: \url{http://trinker.github.io/qdap/vignettes/qdap_vignette.html}.  This}
}
\value{
Places the (1) HTML, (2) source, & (3) R code for the
\emph{Introduction to qdap Vignette} in the user's
\file{R-VERSION/library/qdap/doc}.
}
\description{
Replaces the temporary (place holder) \emph{Introduction to qdap Vignette}
with the actual vignette.
}
\note{
The \pkg{knitr} built HTML approach above takes about 4 minutes.  The
user may choose the faster approach (< 30 seconds) that downloads the HTML
file directly from the Internet (this is for the latest CRAN release of
\pkg{qdap}).  This choice is controlled via the \code{download.html}
argument.  The function will ask for the user's permission before writing the
documents. Once the user has run this function
\code{browseVignettes(package = 'qdap')} will allow access to the new
vignette files.
}
\keyword{vignette}

