% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check_spelling_interactive.check_spelling}
\alias{check_spelling_interactive.check_spelling}
\title{Check Spelling}
\usage{
\method{check_spelling_interactive}{check_spelling}(text.var, click = TRUE,
  ...)
}
\arguments{
\item{text.var}{A \code{\link[qdap]{check_spelling}} object.}

\item{\ldots}{ignored}
}
\description{
View check_spelling check_spelling_interactive.
}
\details{
check_spelling Method for check_spelling_interactive
}

