% Generated by roxygen2 (4.0.1): do not edit by hand
\name{chunker}
\alias{chunker}
\title{Break Text Into Ordered Word Chunks}
\usage{
chunker(text.var, grouping.var = NULL, n.words, n.chunks, as.string = TRUE,
  rm.unequal = FALSE)
}
\arguments{
\item{text.var}{The text variable}

\item{grouping.var}{The grouping variables.  Default \code{NULL} generates
one word list for all text.  Also takes a single grouping variable or a list
of 1 or more grouping variables.}

\item{n.words}{An integer specifying the number of words in each chunk (must
specify n.chunks or n.words).}

\item{n.chunks}{An integer specifying the number of chunks (must specify
n.chunks or n.words).}

\item{as.string}{logical.  If \code{TRUE} the chunks are returned as a single
string.  If \code{FALSE} the chunks are returned as a vector of single words.}

\item{rm.unequal}{logical. If \code{TRUE} final chunks that are unequal in
length to the other chunks are removed.}
}
\value{
Returns a list of text chunks.
}
\description{
Some visualizations and algorithms require text to be broken into chunks of
ordered words.  \code{chunker} breaks text, optionally by grouping
variables, into equal chunks.  The chunk size can be specified by giving
number of words to be in each chunk or the number of chunks.
}
\examples{
with(DATA, chunker(state, n.chunks = 10))
with(DATA, chunker(state, n.words = 10))
with(DATA, chunker(state, n.chunks = 10, as.string=FALSE))
with(DATA, chunker(state, n.chunks = 10, rm.unequal=TRUE))
with(DATA, chunker(state, person, n.chunks = 10))
with(DATA, chunker(state, list(sex, adult), n.words = 10))
with(DATA, chunker(state, person, n.words = 10, rm.unequal=TRUE))

## Bigger data
with(hamlet, chunker(dialogue, person, n.chunks = 10))
with(hamlet, chunker(dialogue, person, n.words = 300))
}
\keyword{chunks}
\keyword{group}
\keyword{text}

