\name{cm_range2long}
\alias{cm_range2long}
\title{Transform Codes to Start-End Durations}
\usage{
  cm_range2long(..., v.name = "variable", list.var = TRUE,
    debug = TRUE)
}
\arguments{
  \item{\ldots}{list object(s) in the form generated by
  cm_time.temp.}

  \item{v.name}{sn optional name for the column created for
  the list.var argument.}

  \item{list.var}{logical.  If TRUE creates a column for
  the data frame created by each time.list passed to
  \code{cm_t2l}.}

  \item{star.end}{logical. If TRUE outputs stop and end
  times for each \code{cm_time.temp} list object.}

  \item{debug}{logical. If TRUE debugging mode is on.
  \code{cm_time2long} will return possible errors in time
  span inputs.}
}
\value{
  Generates a data frame of start and end times for each
  code.
}
\description{
  Transforms the range coding structure(s) from
  cm_range.temp (in list format) into a data frame of start
  and end durations in long format.
}
\examples{
\dontrun{
foo <- list(
    AA = qcv(terms='1'),
    BB = qcv(terms='1:2, 3:10, 19'),
    CC = qcv(terms='1:9, 100:150')
)

foo2  <- list(
    AA = qcv(terms='40'),
    BB = qcv(terms='50:90'),
    CC = qcv(terms='60:90, 100:120, 150'),
    DD = qcv(terms='')
)
dat <- cm_range2long(foo, foo2, v.name = "time")
gantt_wrap(dat, "code", "time")
}
}
\references{
  Miles, M. B. & Huberman, A. M. (1994). An expanded
  sourcebook: Qualitative data analysis. 2nd ed. Thousand
  Oaks, CA: SAGE Publications.
}
\seealso{
  \code{\link{cm_df2long}} \code{\link{cm_time.temp}}
}
\keyword{coding,}
\keyword{span}
\keyword{time}

