% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comma_spacer.R
\name{comma_spacer}
\alias{comma_spacer}
\title{Ensure Space After Comma}
\usage{
comma_spacer(text.var)
}
\arguments{
\item{text.var}{The text variable.}
}
\value{
Returns a vector of strings with commas that have a space after them.
}
\description{
Adds a space after a comma as strip and many other functions may consider a
comma separated string as one word (i.e., \code{"one,two,three"} becomes
\code{"onetwothree"}  rather than \code{"one two three"}).
}
\examples{
\dontrun{
x <- c("the,  dog,went", "I,like,it", "where are you", NA, "why", ",", ",f")
comma_spacer(x)
}
}
\keyword{comma}
\keyword{space}

