% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common.R
\name{common.list}
\alias{common.list}
\title{list Method for common}
\usage{
\method{common}{list}(word.list, overlap = "all", equal.or = "more", ...)
}
\arguments{
\item{word.list}{A list of names character vectors.}

\item{overlap}{Minimum/exact amount of overlap.}

\item{equal.or}{A character vector of c(\code{"equal"}, \code{"greater"},
\code{"more"}, \code{"less"}).}

\item{\dots}{In lieu of word.list the user may input n number of character
vectors.}
}
\description{
list Method for common
}

