\docType{data}
\name{contractions}
\alias{contractions}
\title{Contraction Conversions}
\format{A data frame with 65 rows and 2 variables}
\description{
  A dataset containing common contractions and their
  expanded form.
}
\details{
  \itemize{ \item contraction. The contraction word.  \item
  expanded. The expanded form of the contraction. }
}
\keyword{datasets}

