\name{counts.automated_readability_index}
\alias{counts.automated_readability_index}
\title{Readability Measures}
\usage{
\method{counts}{automated_readability_index}(x, ...)
}
\arguments{
  \item{x}{The automated_readability_index object.}

  \item{\ldots}{ignored automated_readability_index Method
  for counts.}
}
\description{
\code{counts.automated_readability_index} - View counts
from \code{\link[qdap]{automated_readability_index}}.
}

