% Generated by roxygen2 (4.0.2): do not edit by hand
\name{counts.formality}
\alias{counts.formality}
\title{Formality}
\usage{
\method{counts}{formality}(x, ...)
}
\arguments{
\item{x}{The \code{\link[qdap]{formality}} object.}

\item{\ldots}{ignored}
}
\description{
View formality counts.
}
\details{
formality Method for counts
}

