% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated_readability_index.R
\name{counts.fry}
\alias{counts.fry}
\title{Readability Measures}
\usage{
\method{counts}{fry}(x, ...)
}
\arguments{
\item{x}{The fry object.}

\item{\ldots}{ignored}
}
\description{
\code{counts.fry} - View counts from \code{\link[qdap]{fry}}.
}
\details{
fry Method for counts.
}
