% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/object_pronoun_type.R
\name{counts.object_pronoun_type}
\alias{counts.object_pronoun_type}
\title{Question Counts}
\usage{
\method{counts}{object_pronoun_type}(x, ...)
}
\arguments{
\item{x}{The \code{\link[qdap]{object_pronoun_type}} object.}

\item{\ldots}{ignored}
}
\description{
View object_pronoun_type counts.
}
\details{
object_pronoun_type Method for counts
}

