% Generated by roxygen2 (4.0.0): do not edit by hand
\name{counts.polarity}
\alias{counts.polarity}
\title{Polarity}
\usage{
\method{counts}{polarity}(x, ...)
}
\arguments{
\item{x}{The polarity object.}

\item{\ldots}{ignored}
}
\description{
\code{counts.polarity} - View counts from \code{\link[qdap]{polarity}}.
}
\details{
polarity Method for counts.
}

