% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pronoun_type.R
\name{counts.pronoun_type}
\alias{counts.pronoun_type}
\title{Question Counts}
\usage{
\method{counts}{pronoun_type}(x, ...)
}
\arguments{
\item{x}{The \code{\link[qdap]{pronoun_type}} object.}

\item{\ldots}{ignored}
}
\description{
View pronoun_type counts.
}
\details{
pronoun_type Method for counts
}

