% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/question_type.R
\name{counts.question_type}
\alias{counts.question_type}
\title{Question Counts}
\usage{
\method{counts}{question_type}(x, ...)
}
\arguments{
\item{x}{The \code{\link[qdap]{question_type}} object.}

\item{\ldots}{ignored}
}
\description{
View question_type counts.
}
\details{
question_type Method for counts
}

