% Generated by roxygen2 (4.0.1): do not edit by hand
\name{counts.SMOG}
\alias{counts.SMOG}
\title{Readability Measures}
\usage{
\method{counts}{SMOG}(x, ...)
}
\arguments{
\item{x}{The SMOG object.}

\item{\ldots}{ignored}
}
\description{
\code{counts.SMOG} - View counts from \code{\link[qdap]{SMOG}}.
}
\details{
SMOG Method for counts.
}

