% Generated by roxygen2 (4.0.1): do not edit by hand
\name{counts.termco}
\alias{counts.termco}
\title{Term Counts}
\usage{
\method{counts}{termco}(x, ...)
}
\arguments{
\item{x}{The \code{\link[qdap]{termco}} object.}

\item{\ldots}{ignored}
}
\description{
View termco counts.
}
\details{
termco Method for counts
}

