% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/word_length.R
\name{counts.word_length}
\alias{counts.word_length}
\title{Word Length Counts}
\usage{
\method{counts}{word_length}(x, ...)
}
\arguments{
\item{x}{The \code{\link[qdap]{word_length}} object.}

\item{\ldots}{ignored}
}
\description{
View word_length counts.
}
\details{
word_length Method for counts
}

