% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/word_stats.R
\name{counts.word_stats}
\alias{counts.word_stats}
\title{Word Stats}
\usage{
\method{counts}{word_stats}(x, ...)
}
\arguments{
\item{x}{The \code{\link[qdap]{word_stats}} object.}

\item{\ldots}{ignored}
}
\description{
View word_stats counts.
}
\details{
word_stats Method for counts
}

