% Generated by roxygen2 (4.0.0): do not edit by hand
\name{cumulative}
\alias{cumulative}
\alias{cumulative.combo_syllable_sum}
\alias{cumulative.end_mark}
\alias{cumulative.formality}
\alias{cumulative.polarity}
\alias{cumulative.pos}
\alias{cumulative.pos_by}
\alias{cumulative.syllable_freq}
\title{Cumulative Scores}
\usage{
cumulative(x, ...)

\method{cumulative}{end_mark}(x, ...)

\method{cumulative}{formality}(x, ...)

\method{cumulative}{pos}(x, ...)

\method{cumulative}{pos_by}(x, ...)

\method{cumulative}{polarity}(x, ...)

\method{cumulative}{syllable_freq}(x, ...)

\method{cumulative}{combo_syllable_sum}(x, ...)
}
\arguments{
\item{x}{A qdap object with an accompanying \code{cumulative} method.}

\item{\ldots}{ignored}
}
\description{
\code{cumulative} - Generate rolling/cumulative scores for select \pkg{qdap}
objects.

\code{cumulative.end_mark} - Generate end_mark over time (duration in
sentences).

\code{cumulative.formality} - Generate formality over time (duration in
sentences).

\code{cumulative.pos} - Generate formality over time (duration in
sentences).

\code{cumulative.pos_by} - Generate formality over time (duration in
sentences).

\code{cumulative.polarity} - Generate polarity over time (duration in
sentences).

\code{cumulative.syllable_freq} - Generate syllable_freq over time (duration in
sentences).

\code{cumulative.combo_syllable_sum} - Generate combo_syllable_sum over time (duration in
sentences).
}

