% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qdap-package.R
\docType{data}
\name{DATA.SPLIT}
\alias{DATA.SPLIT}
\title{Fictitious Split Sentence Classroom Dialogue}
\format{A data frame with 15 rows and 8 variables}
\usage{
data(DATA.SPLIT)
}
\description{
A \code{\link[qdap]{sentSplit}} version of the \code{\link[qdap]{DATA}}
dataset.
}
\details{
\itemize{
  \item person. Speaker
  \item tot. Turn of talk with sub sentences
  \item TOT. Turn of talk
  \item sex. Gender
  \item adult. Dummy coded adult (0-no; 1-yes)
  \item code. Dialogue coding scheme
  \item state. Statement (dialogue)
  \item stem.text. A stemmed version of the text.var
}
}
\keyword{datasets}

