% Generated by roxygen2 (4.0.0): do not edit by hand
\name{df2tm_corpus}
\alias{df2tm_corpus}
\alias{tm2qdap}
\title{Deprecated qdap Functions}
\usage{
df2tm_corpus(text.var, grouping.var = NULL, demographic.vars, ...)

tm2qdap(x)
}
\arguments{
\item{text.var}{The text variable or a \code{\link[qdap]{wfm}} object.}

\item{grouping.var}{The grouping variables.  Default \code{NULL} generates
one word list for all text.  Also takes a single grouping variable or a list
of 1 or more grouping variables.}

\item{demographic.vars}{Additional demographic information about the grouping
variables.  This is a data.frame, list of equal length vectors, or a single
vector corresponding to the grouping variable/text variable.  This
information will be mapped to the DMetaData in the \code{\link[tm]{Corpus}}.}

\item{\ldots}{Other arguments passed to \code{\link[qdap]{sentSplit}}.}

\item{x}{A \code{\link[tm]{TermDocumentMatrix}}/\code{\link[tm]{DocumentTermMatrix}}.}
}
\value{
\code{df2tm_corpus} - Converts a qdap oriented dataframe and returns
a \code{\link[tm]{Corpus}}.

\code{tm2qdap} - Returns a \code{\link[qdap]{wfm}} object or
\code{weight} object.
}
\description{
\code{df2tm_corpus} - Convert a qdap dataframe to a tm package
\code{\link[tm]{Corpus}}.

\code{tm2qdap} - Convert the \pkg{tm} package's
\code{\link[tm]{TermDocumentMatrix}}/\code{\link[tm]{DocumentTermMatrix}} to
\code{\link[qdap]{wfm}}.
}
\section{Warning}{
 \code{df2tm_corpus} - function is deprecated.  It will be
removed in a subsequent version of qdap.  Use \code{as.Corpus} instead.
}

\section{Warning}{
 \code{tm2qdap} - function is deprecated.  It will be
removed in a subsequent version of qdap.  Use \code{as.wfm} instead.
}

