\docType{data}
\name{DICTIONARY}
\alias{DICTIONARY}
\title{Nettalk Corpus Syllable Data Set}
\format{A data frame with 20137 rows and 2 variables}
\description{
  A dataset containing syllable counts.
}
\details{
  \itemize{ \item word. The word \item syllables. Number of
  syllables }
}
\note{
  This data set is based on the Nettalk Corpus but has some
  researcher word deletions and additions based on the
  needs of the \code{\link[qdap]{syllable.sum}} algorithm.
}
\references{
  Sejnowski, T.J., and Rosenberg, C.R. (1987). "Parallel
  networks that learn to pronounce English text" in Complex
  Systems, 1, 145-168. Retrieved from:
  \url{http://archive.ics.uci.edu/ml/datasets/Connectionist+Bench+(Nettalk+Corpus)}

  \href{http://archive.ics.uci.edu/ml/machine-learning-databases/undocumented/connectionist-bench/nettalk/}{UCI
  Machine Learning Repository website}
}
\keyword{datasets}

