\docType{data}
\name{emoticon}
\alias{emoticon}
\title{Emoticons Data Set}
\format{A data frame with 81 rows and 2 variables}
\description{
  A dataset containing common emoticons (adapted from
  \href{http://www.lingo2word.com/lists/emoticon_listH.html}{Popular
  Emoticon List}).
}
\details{
  \itemize{ \item meaning. The meaning of the emoticon
  \item emoticon. The graphic representation of the
  emoticon }
}
\references{
  \url{http://www.lingo2word.com/lists/emoticon_listH.html}
}
\keyword{datasets}

