\docType{data}
\name{env.syl}
\alias{env.syl}
\title{Syllable Lookup Environment}
\format{A environment with the DICTIONARY data set.}
\description{
  A dataset containing a syllable lookup environment (see
  \code{link[qdap]{DICTIONARY}}).
}
\details{
  For internal use.
}
\references{
  \href{http://archive.ics.uci.edu/ml/machine-learning-databases/undocumented/connectionist-bench/nettalk/}{UCI
  Machine Learning Repository website}
}
\keyword{datasets}

