\docType{data}
\name{interjections}
\alias{interjections}
\title{Interjections}
\format{A character vector with 139 elements}
\description{
  A dataset containing a character vector of common
  interjections.
}
\references{
  \url{http://www.vidarholen.net/contents/interjections/}
}
\keyword{datasets}

