% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lexical_classification}
\alias{lexical_classification}
\title{Lexical Classification Score}
\usage{
lexical_classification(text.var, grouping.var = NULL,
  order.by.lexical_classification = TRUE,
  function.words = qdapDictionaries::function.words, bracket = "all", ...)
}
\arguments{
\item{text.var}{The text variable.}

\item{grouping.var}{The grouping variables.  Default \code{NULL} generates
one word list for all text.  Also takes a single grouping variable or a list
of 1 or more grouping variables.}

\item{order.by.lexical_classification}{logical.  If \code{TRUE} orders the
results by #' lexical_classification score.}

\item{function.words}{A vector of function words.  Default is
\code{\link[qdapDictionaries]{function.words}}.}

\item{bracket}{The bracket type to remove.  Use \code{NULL} to not remove
bracketed substrings.  See \code{bracket} argument in
\code{\link[qdap]{bracketX}} for bracket types.}

\item{\ldots}{Other arguments passed to \code{\link[qdap]{bracketX}}.}
}
\value{
A list containing at the following components:
\item{content}{A \code{data.frame} of all content words used and corresponding frequencies}
\item{functional}{A \code{data.frame} of all content words used and corresponding frequencies}
\item{raw}{Sentence level descriptive statistics on content vs. functional word use}
\item{lexical_classification}{Summarized (grouping variable level) descriptive statistics for content vs. functional word use}
}
\description{
Transcript apply lexical classification score (content to functional word
proportion) by grouping variable(s) and optionally plot
the breakdown of the model.
}
\details{
Content words (i.e., nouns, verbs, adjectives, and adverbs) tend to
be the words speakers stresses in language use.  Whereas, functional words
are the "glue" that holds the content together.  Speakers devote much less
time and stress to these words (i.e., pronouns, articles, conjunctions,
quantifiers, and prepositions).
}
\examples{
\dontrun{
lexical_classification("I did not like the dog.")
lexical_classification(DATA.SPLIT$state, DATA.SPLIT$person)

(out <- with(pres_debates2012, lexical_classification(dialogue, list(person, time))))
plot(out)

scores(out)

out2 <- preprocessed(out)
htruncdf(out2)
plot(out2)

plot(out[["content"]])
dev.new()
plot(out[["functional"]])

## cloud of functional vs. content
## Highlight Content Words
set.seed(10)
par(mar = c(0,0,0,0))
list(
        content = out[["content"]],
        functional = out[["functional"]]
    ) \%>\%
    list_df2df("type") \%>\%
    dplyr::mutate(colors = ifelse(type == "functional", "gray80", "blue")) \%>\%
    with(., wordcloud::wordcloud(
        word,
        freq,
        min.freq = 8,
        random.order=FALSE,
        ordered.colors = TRUE,
        colors = colors
    ))
mtext("2012 Presidential Debates:\\nFunctional vs. Content Word Use", padj=1.25)
legend(
    .05, .12, bty = "n",
    legend = c("functional", "content"),
    fill = c("gray80", "blue"),
    cex = .7
)

## Highlight Functional Words
set.seed(10)
par(mar = c(0,0,0,0))
list(
        content = out[["content"]],
        functional = out[["functional"]]
    ) \%>\%
    list_df2df("type") \%>\%
    dplyr::mutate(colors = ifelse(type == "functional", "red", "gray80")) \%>\%
    with(., wordcloud::wordcloud(
        word,
        freq,
        min.freq = 8,
        random.order=FALSE,
        ordered.colors = TRUE,
        colors = colors
    ))
mtext("2012 Presidential Debates:\\nFunctional vs. Content Word Use", padj=1.25)
legend(
    .05, .12, bty = "n",
    legend = c("functional", "content"),
    fill = c("red", "gray80"),
    cex = .7
)
}
}
\references{
Chung, C. & Pennebaker, J. (2007). The Psychological Functions of Function Words. In K. Fiedler (Ed.) Social Communication (pp. 343-359). New York: Psychology Press.

Pulvermuller, F. (1999). Words in the brain's language. Behavioral and Brain Sciences, 22, pp. 253-279. doi:10.1017/S0140525X9900182X

 Segalowitz, S. J. & Lane, K. (2004). Perceptual fluency and lexical access for function versus content words. Behavioral and Brain Sciences, 27, 307-308. doi:10.1017/S0140525X04310071

Bell, A., Brenier, J. M., Gregory, M., Girand, C. & Jurafsky, D. (2009).  Predictability Effects on Durations of Content and Function Words in Conversational English.  Journal of Memory and Language, 60(1), 92-111. doi:10.1016/j.jml.2008.06.003
}
\keyword{content}
\keyword{functional,}
\keyword{lexical_classification,}
\keyword{parts-of-speech,}

