\docType{data}
\name{NAMES}
\alias{NAMES}
\title{First Names and Gender (U.S.)}
\format{A data frame with 5494 rows and 7 variables}
\description{
  A dataset containing 1990 U.S. census data on first
  names.
}
\details{
  \itemize{ \item name. A first name.  \item per.freq.
  Frequency in percent of the name by gender.  \item
  cum.freq. Cumulative frequency in percent of the name by
  gender.  \item rank. Rank of the name by gender.  \item
  gender. Gender of the combined male/female list (M/F).
  \item gender2. Gender of the combined male/female list
  with "B" in place of overlapping (M/F) names.  \item
  pred.sex. Predicted gender of the names with B's in
  \code{gender2} repalced with the gender that had a higher
  \code{per.freq}. }
}
\references{
  \url{http://www.census.gov/genealogy/www/data/1990surnames/names_files.html}
}
\keyword{datasets}

