\docType{data}
\name{NAMES_LIST}
\alias{NAMES_LIST}
\title{First Names and Predictive Gender (U.S.) List}
\format{A list with 26 elements}
\description{
  A list version of the \code{link[qdap]{NAMES_SEX}})
  dataset broken down by first letter.
}
\details{
  Alphabetical list of dataframes with the following
  variables: \itemize{ \item name. A first name.  \item
  gender2. Gender of the combined male/female list with "B"
  in place of overlapping (M/F) names.  \item pred.sex.
  Predicted gender of the names with B's in \code{gender2}
  repalced with the gender that had a higher
  \code{per.freq}. }
}
\references{
  \url{http://www.census.gov/genealogy/www/data/1990surnames/names_files.html}
}
\keyword{datasets}

