% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formality.R
\name{Network.formality}
\alias{Network.formality}
\title{Network Formality}
\usage{
\method{Network}{formality}(
  x,
  contextual = "yellow",
  formal = "red",
  edge.constant,
  title = NULL,
  digits = 3,
  plus.300.color = "grey40",
  under.300.color = "grey88",
  missing.color = "purple",
  ...
)
}
\arguments{
\item{x}{A \code{\link[qdap]{formality}} object.}

\item{contextual}{The color to use for 0\% formality (purely contextual).}

\item{formal}{The color to use for 100\% formality (purely formal).}

\item{edge.constant}{A constant to multiple edge width by.}

\item{title}{The title to apply to the \code{Network}ed image(s).}

\item{digits}{The number of digits to use in the current turn of talk 
formality.}

\item{plus.300.color}{The bar color to use for grouping variables exceeding 
299 words per Heylighen & Dewaele's (2002) minimum word recommendations.}

\item{under.300.color}{The bar color to use for grouping variables less 
than 300 words per Heylighen & Dewaele's (2002) minimum word recommendations.}

\item{missing.color}{The color to use in a network plot for edges 
corresponding to missing text data.  Use \code{\link[stats]{na.omit}} before 
hand to remove the missing values all together.}

\item{\ldots}{Other arguments passed to \code{\link[qdap]{discourse_map}}.}
}
\description{
\code{Network.formality} - Network a \code{\link[qdap]{formality}} object.
}
\details{
formality Method for Network
}
