% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lexical_classification.R
\name{Network.lexical_classification}
\alias{Network.lexical_classification}
\title{Network Lexical Classification}
\usage{
\method{Network}{lexical_classification}(x, functional = "yellow",
  content = "red", edge.constant, title = NULL, digits = 2, ...)
}
\arguments{
\item{x}{A \code{\link[qdap]{lexical_classification}} object.}

\item{functional}{The color to use for 0\% lexical_classification (purely
functional).}

\item{content}{The color to use for 100\% lexical_classification (purely
content).}

\item{edge.constant}{A constant to multiple edge width by.}

\item{title}{The title to apply to the Networked image(s).}

\item{digits}{The number of digits to use in the current turn of talk
lexical_classification.}

\item{\ldots}{Other arguments passed to \code{\link[qdap]{discourse_map}}.}
}
\description{
\code{Network.lexical_classification} - Network a
\code{\link[qdap]{lexical_classification}} object.
}
\details{
lexical_classification Method for Network
}

