% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polarity.R
\name{Network.polarity}
\alias{Network.polarity}
\title{Network Polarity}
\usage{
\method{Network}{polarity}(x, negative = "blue", positive = "red",
  neutral = "yellow", edge.constant, title = NULL, digits = 3, ...)
}
\arguments{
\item{x}{A \code{\link[qdap]{polarity}} object.}

\item{negative}{The color to use for negative polarity.}

\item{positive}{The color to use for positive polarity.}

\item{neutral}{The color to use for neutral polarity.}

\item{edge.constant}{A constant to multiple edge width by.}

\item{title}{The title to apply to the Networked image(s).}

\item{digits}{The number of digits to use in the current turn of talk
polarity.}

\item{\ldots}{Other arguments passed to \code{\link[qdap]{discourse_map}}.}
}
\description{
\code{Network.polarity} - Network a \code{\link[qdap]{polarity}} object.
}
\details{
polarity Method for Network
}

