% Generated by roxygen2 (4.0.0): do not edit by hand
\name{Network}
\alias{Network}
\title{Generic Network Method}
\usage{
Network(x, ...)
}
\arguments{
\item{x}{A select qdap object.}

\item{\ldots}{Arguments passed to Network method of other classes.}
}
\value{
Returns a network plot.
}
\description{
Create a network plt for select qdap outputs.
}

