\docType{data}
\name{OnixTxtRetToolkitSWL1}
\alias{OnixTxtRetToolkitSWL1}
\title{Onix Text Retrieval Toolkit Stopword List 1}
\format{A character vector with 404 elements}
\description{
  A stopword list containing a character vector of
  stopwords.
}
\details{
  \href{http://www.lextek.com/manuals/onix/stopwords1.html}{From
  Onix Text Retrieval Toolkit API Reference}: "This
  stopword list is probably the most widely used stopword
  list. It covers a wide number of stopwords without
  getting too aggressive and including too many words which
  a user might search upon."
}
\note{
  Reduced from the original 429 words to 404.
}
\references{
  \url{http://www.lextek.com/manuals/onix/stopwords1.html}
}
\keyword{datasets}

