% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Animate.character.R
\name{plot.animated_character}
\alias{plot.animated_character}
\title{Plots an animated_character  Object}
\usage{
\method{plot}{animated_character}(x, ...)
}
\arguments{
\item{x}{The animated_character  object.}

\item{\ldots}{Other arguments passed to \code{print.animated_character}.}
}
\description{
Plots an animated_character  object.
}

