% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.animated_discourse_map}
\alias{plot.animated_discourse_map}
\title{Plots a animated_discourse_map  Object}
\usage{
\method{plot}{animated_discourse_map}(x, ...)
}
\arguments{
\item{x}{The animated_discourse_map  object.}

\item{\ldots}{Other arguments passed to \code{print.animated_discourse_map }.}
}
\description{
Plots a animated_discourse_map  object.
}

