% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.animated_lexical_classification}
\alias{plot.animated_lexical_classification}
\title{Plots a animated_lexical_classification  Object}
\usage{
\method{plot}{animated_lexical_classification}(x, ...)
}
\arguments{
\item{x}{The animated_lexical_classification  object.}

\item{\ldots}{Other arguments passed to \code{print.animated_lexical_classification }.}
}
\description{
Plots a animated_lexical_classification  object.
}

