\name{plot.automated_readability_index}
\alias{plot.automated_readability_index}
\title{Plots a automated_readability_index Object}
\usage{
\method{plot}{automated_readability_index}(x, ...)
}
\arguments{
  \item{x}{The automated_readability_index object.}

  \item{\ldots}{ignored}
}
\description{
Plots a automated_readability_index object.
}

