\name{plot.character.table}
\alias{plot.character.table}
\title{Plots a character.table Object}
\usage{
  \method{plot}{character.table} (x, label = FALSE, ...)
}
\arguments{
  \item{x}{The character.table object}

  \item{label}{logical.  If TRUE the cells of the heat map
  plot will be labeled with raw and proportional values.}

  \item{\ldots}{Other arguments passed to qheat}
}
\description{
  Plots a character.table object.
}

