% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/word_count.R
\name{plot.character_table}
\alias{plot.character_table}
\title{Plots a character_table Object}
\usage{
\method{plot}{character_table}(x, label = FALSE, lab.digits = 1,
  percent = NULL, zero.replace = NULL, ...)
}
\arguments{
\item{x}{The character_table  object}

\item{label}{logical.  If \code{TRUE} the cells of the heat map plot will be
labeled with count and proportional values.}

\item{lab.digits}{Integer values specifying the number of digits to be
printed if \code{label} is \code{TRUE}.}

\item{percent}{logical.  If \code{TRUE} output given as percent.  If
\code{FALSE} the output is proportion.  If \code{NULL} uses the value from
\code{\link[qdap]{question_type}}.  Only used if \code{label} is \code{TRUE}.}

\item{zero.replace}{Value to replace 0 values with.  If \code{NULL} uses the
value from \code{\link[qdap]{question_type}}.  Only used if \code{label} is
\code{TRUE}.}

\item{\ldots}{Other arguments passed to \code{\link[qdap]{qheat}}}
}
\description{
Plots a character_table  object.
}

