% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.combo_syllable_sum}
\alias{plot.combo_syllable_sum}
\title{Plots a combo_syllable_sum Object}
\usage{
\method{plot}{combo_syllable_sum}(x, ...)
}
\arguments{
\item{x}{The combo_syllable_sum object.}

\item{\ldots}{ignored}
}
\description{
Plots a combo_syllable_sum object.
}

