% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.cumulative_animated_lexical_classification}
\alias{plot.cumulative_animated_lexical_classification}
\title{Plots a cumulative_animated_lexical_classification Object}
\usage{
\method{plot}{cumulative_animated_lexical_classification}(x, ...)
}
\arguments{
\item{x}{The cumulative_animated_lexical_classification object.}

\item{\ldots}{ignored}
}
\description{
Plots a cumulative_animated_lexical_classification object.
}

