% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/syllable_sum.R
\name{plot.cumulative_combo_syllable_sum}
\alias{plot.cumulative_combo_syllable_sum}
\title{Plots a cumulative_combo_syllable_sum Object}
\usage{
\method{plot}{cumulative_combo_syllable_sum}(x, ...)
}
\arguments{
\item{x}{The cumulative_combo_syllable_sum object.}

\item{\ldots}{ignored}
}
\description{
Plots a cumulative_combo_syllable_sum object.
}

