% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.cumulative_end_mark}
\alias{plot.cumulative_end_mark}
\title{Plots a cumulative_end_mark Object}
\usage{
\method{plot}{cumulative_end_mark}(x, ...)
}
\arguments{
\item{x}{The cumulative_end_mark object.}

\item{\ldots}{ignored}
}
\description{
Plots a cumulative_end_mark object.
}

