% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formality.R
\name{plot.cumulative_formality}
\alias{plot.cumulative_formality}
\title{Plots a cumulative_formality Object}
\usage{
\method{plot}{cumulative_formality}(x, ...)
}
\arguments{
\item{x}{The cumulative_formality object.}

\item{\ldots}{ignored}
}
\description{
Plots a cumulative_formality object.
}
