% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lexical_classification.R
\name{plot.cumulative_lexical_classification}
\alias{plot.cumulative_lexical_classification}
\title{Plots a cumulative_lexical_classification Object}
\usage{
\method{plot}{cumulative_lexical_classification}(x, ...)
}
\arguments{
\item{x}{The cumulative_lexical_classification object.}

\item{\ldots}{ignored}
}
\description{
Plots a cumulative_lexical_classification object.
}

