% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.cumulative_polarity}
\alias{plot.cumulative_polarity}
\title{Plots a cumulative_polarity Object}
\usage{
\method{plot}{cumulative_polarity}(x, ...)
}
\arguments{
\item{x}{The cumulative_polarity object.}

\item{\ldots}{ignored}
}
\description{
Plots a cumulative_polarity object.
}

