% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plot.cumulative_syllable_freq}
\alias{plot.cumulative_syllable_freq}
\title{Plots a cumulative_syllable_freqObject}
\usage{
\method{plot}{cumulative_syllable_freq}(x, ...)
}
\arguments{
\item{x}{The cumulative_syllable_freqobject.}

\item{\ldots}{ignored}
}
\description{
Plots a cumulative_syllable_freqobject.
}

