% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversity.R
\name{plot.diversity}
\alias{plot.diversity}
\title{Plots a diversity object}
\usage{
\method{plot}{diversity}(x, ...)
}
\arguments{
\item{x}{The diversity object}

\item{\ldots}{Other arguments passed to \code{qheat}}
}
\description{
Plots a diversity object.
}

