% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/end_mark.R
\name{plot.end_mark}
\alias{plot.end_mark}
\title{Plots an end_mark Object}
\usage{
\method{plot}{end_mark}(x, ...)
}
\arguments{
\item{x}{The end_mark object.}

\item{\ldots}{ignored}
}
\description{
Plots an end_mark object.
}

