% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/end_mark.R
\name{plot.end_mark_by}
\alias{plot.end_mark_by}
\title{Plots a end_mark_by Object}
\usage{
\method{plot}{end_mark_by}(x, values = FALSE, ...)
}
\arguments{
\item{x}{The end_mark_by object.}

\item{values}{logical.  If \code{TRUE} the cell values will be included on
the heatmap.}

\item{\ldots}{Other arguments passed to \code{\link[qdap]{qheat}}.}
}
\description{
Plots a end_mark_by object.
}

