% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.flesch_kincaid}
\alias{plot.flesch_kincaid}
\title{Plots a flesch_kincaid Object}
\usage{
\method{plot}{flesch_kincaid}(x, ...)
}
\arguments{
\item{x}{The readability_score object.}

\item{\ldots}{ignored}
}
\description{
Plots a flesch_kincaid object.
}

