\name{plot.formality_preprocessed}
\alias{plot.formality_preprocessed}
\title{Plots a formality_preprocessed Object}
\usage{
\method{plot}{formality_preprocessed}(x, ...)
}
\arguments{
  \item{x}{The formality_preprocessed object.}

  \item{\ldots}{ignored}
}
\description{
Plots a formality_preprocessed object.
}

