% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kullback_leibler.R
\name{plot.kullback_leibler}
\alias{plot.kullback_leibler}
\title{Plots a kullback_leibler object}
\usage{
\method{plot}{kullback_leibler}(x, digits = 3, ...)
}
\arguments{
\item{x}{The kullback_leibler object}

\item{digits}{Number of decimal places to print.}

\item{\ldots}{Other arguments passed to \code{qheat}}
}
\description{
Plots a kullback_leibler object.
}

