% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lexical_classification.R
\name{plot.lexical}
\alias{plot.lexical}
\title{Plots a lexical Object}
\usage{
\method{plot}{lexical}(x, min.freq = 1, rot.per = 0, random.order = FALSE,
  title = TRUE, title.color = "blue", ...)
}
\arguments{
\item{x}{The lexical object.}

\item{min.freq}{Words with frequency below \code{min.freq} will not be plotted.}

\item{rot.per}{Proportion words with 90 degree rotation.}

\item{random.order}{logical.  If code{TRUE} plot words in random order. If \code{FALSE}, they will be plotted in decreasing frequency.}

\item{title}{The title of the plot.  Use \code{NULL} to eliminate.}

\item{title.color}{The color of the title.}

\item{\ldots}{Other arguments passed to \code{\link[wordcloud]{wordcloud}}.}
}
\description{
Plots a lexical object.
}

