% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lexical_classification.R
\name{plot.lexical_classification_preprocessed}
\alias{plot.lexical_classification_preprocessed}
\title{Plots a lexical_classification_preprocessed Object}
\usage{
\method{plot}{lexical_classification_preprocessed}(x, jitter = 0.1,
  text.size = 3.5, alpha = 0.3, ncol = 3, ...)
}
\arguments{
\item{x}{The lexical_classification_preprocessed object.}

\item{jitter}{The amount to jitter the points by in the bocplots.}

\item{text.size}{The text size to use for plotting the mean in the boxplots.}

\item{alpha}{The alpha level to use for points.}

\item{ncol}{The number of columns to use for \code{\link[ggplot2]{facet_wrap}}.}

\item{\ldots}{ignored}
}
\description{
Plots a lexical_classification_preprocessed object.
}

