% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plot.Network}
\alias{plot.Network}
\title{Plots a Network  Object}
\usage{
\method{plot}{Network}(x, ...)
}
\arguments{
\item{x}{The Network  object.}

\item{\ldots}{Other arguments passed to \code{print.Network }.}
}
\description{
Plots a Network  object.
}

