% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polarity.R
\name{plot.polarity_score}
\alias{plot.polarity_score}
\title{Plots a polarity_score Object}
\usage{
\method{plot}{polarity_score}(x, error.bar.height = 0.35,
  error.bar.size = 0.5, error.bar.alpha = 0.3, ...)
}
\arguments{
\item{x}{The polarity_score object.}

\item{error.bar.height}{The height of the error bar ends.}

\item{error.bar.size}{The size/thickness of the error bars.}

\item{error.bar.alpha}{The alpha level of the error bars.}

\item{\ldots}{ignored}
}
\description{
Plots a polarity_score object.
}

