\name{plot.question_type}
\alias{plot.question_type}
\title{Plots a question_type Object}
\usage{
  \method{plot}{question_type} (x, label = FALSE, ...)
}
\arguments{
  \item{x}{The question_type object.}

  \item{label}{logical.  If TRUE the cells of the heat map
  plot will be labeled with raw and proportional values.}

  \item{\ldots}{Other arguments passed to qheat.}
}
\description{
  Plots a question_type object.
}

