\name{plot.readability_score}
\alias{plot.readability_score}
\title{Plots a readability_score Object}
\usage{
\method{plot}{readability_score}(x, alpha = 0.3, ...)
}
\arguments{
  \item{x}{The readability_score object.}

  \item{alpha}{The alpha level to be used for the points.}

  \item{\ldots}{ignored}
}
\description{
Plots a readability_score object.
}

