% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/automated_readability_index.R
\name{plot.SMOG}
\alias{plot.SMOG}
\title{Plots a SMOG Object}
\usage{
\method{plot}{SMOG}(x, ...)
}
\arguments{
\item{x}{The readability_score object.}

\item{\ldots}{ignored}
}
\description{
Plots a SMOG object.
}

