% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subject_pronoun_type.R
\name{plot.subject_pronoun_type}
\alias{plot.subject_pronoun_type}
\title{Plots an subject_pronoun_type Object}
\usage{
\method{plot}{subject_pronoun_type}(x, type = 1, ...)
}
\arguments{
\item{x}{The subject_pronoun_type object.}

\item{type}{An integer of \code{1}, \code{2}, \code{3}) corresponding to
1 - heat map; 2 - lexical dispersion plot; 3 - facetted bar graph.}

\item{\ldots}{Other arguments passed to \code{\link[qdap]{qheat}},
\code{\link[qdap]{dispersion_plot}}, or \code{\link[ggplot2]{facet_wrap}}.}
}
\description{
Plots an subject_pronoun_type object.
}

