% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plot.syllable_freq}
\alias{plot.syllable_freq}
\title{Plots a syllable_freq Object}
\usage{
\method{plot}{syllable_freq}(x, ...)
}
\arguments{
\item{x}{The syllable_freq object.}

\item{\ldots}{ignored}
}
\description{
Plots a syllable_freq object.
}

