\name{plot.table_count}
\alias{plot.table_count}
\title{Plots a table_count Object}
\usage{
\method{plot}{table_count}(x, values = TRUE, high = "red", ...)
}
\arguments{
  \item{x}{The table_count object.}

  \item{high}{The color to be used for higher values.}

  \item{values}{logical.  If \code{TRUE} the cell values
  will be included on the heatmap.}

  \item{\ldots}{Arguments passed to qheat.}
}
\description{
Plots a table_count object.
}

