\name{plot.table_proportion}
\alias{plot.table_proportion}
\title{Plots a table_proportion Object}
\usage{
\method{plot}{table_proportion}(x, values = TRUE, high = "red", ...)
}
\arguments{
  \item{x}{The table_proportion object.}

  \item{high}{The color to be used for higher values.}

  \item{values}{logical.  If \code{TRUE} the cell values
  will be included on the heatmap.}

  \item{\ldots}{Other arguments passed to
  \code{\link[qdap]{qheat}}.}
}
\description{
Plots a table_proportion object.
}

