% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/termco.R
\name{plot.termco}
\alias{plot.termco}
\title{Plots a termco object}
\usage{
\method{plot}{termco}(x, label = FALSE, lab.digits = 1, percent = NULL,
  zero.replace = NULL, ...)
}
\arguments{
\item{x}{The termco object.}

\item{label}{logical.  If TRUE the cells of the heat map plot will be labeled
with count and proportional values.}

\item{lab.digits}{Integer values specifying the number of digits to be
printed if \code{label} is TRUE.}

\item{percent}{logical.  If TRUE output given as percent.  If FALSE the
output is proportion.  If NULL uses the value from
\code{\link[qdap]{termco}}.  Only used if \code{label} is TRUE.}

\item{zero.replace}{Value to replace 0 values with.  If NULL uses the value
from \code{\link[qdap]{termco}}.  Only used if \code{label} is TRUE.}

\item{\ldots}{Other arguments passed to qheat.}
}
\description{
Plots a termco object.
}

