% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/type_token_ratio.R
\name{plot.type_token_ratio}
\alias{plot.type_token_ratio}
\title{Plots a type_token_ratio Object}
\usage{
\method{plot}{type_token_ratio}(x, ...)
}
\arguments{
\item{x}{The type_token_ratio object.}

\item{\ldots}{ignored.}
}
\description{
Plots a type_token_ratio object.
}

