% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wfm.R
\name{plot.weighted_wfm}
\alias{plot.weighted_wfm}
\title{Plots a weighted_wfm object}
\usage{
\method{plot}{weighted_wfm}(x, non.zero = FALSE, digits = 0,
  by.column = NULL, high = ifelse(non.zero, "black", "blue"),
  grid = ifelse(non.zero, "black", "white"), plot = TRUE, ...)
}
\arguments{
\item{x}{The weighted_wfm object}

\item{non.zero}{logical.  If \code{TRUE} all values converted to dummy coded
based on x_ij > 0.}

\item{digits}{The number of digits displayed if \code{values} is \code{TRUE}.}

\item{by.column}{logical.  If \code{TRUE} applies scaling to the column.  If
\code{FALSE}  applies scaling by row (use \code{NULL} to turn off scaling).}

\item{high}{The color to be used for higher values.}

\item{grid}{The color of the grid (Use \code{NULL} to remove the grid).}

\item{plot}{logical.  If \code{TRUE} the plot will automatically plot.
The user may wish to set to \code{FALSE} for use in knitr, sweave, etc.
to add additional plot layers.}

\item{\ldots}{Other arguments passed to qheat.}
}
\description{
Plots a weighted_wfm object.
}

