% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wfm.R
\name{plot.wfdf}
\alias{plot.wfdf}
\title{Plots a wfdf object}
\usage{
\method{plot}{wfdf}(x, ...)
}
\arguments{
\item{x}{The wfdf object}

\item{\ldots}{Other arguments passed to \code{\link[qdap]{plot.wfm}}.}
}
\description{
Plots a wfdf object.
}

