% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/word_position.R
\name{plot.word_position}
\alias{plot.word_position}
\title{Plots a word_position object}
\usage{
\method{plot}{word_position}(x, qheat = TRUE, scale = TRUE, ...)
}
\arguments{
\item{x}{The word_position object.}

\item{qheat}{logical.  If \code{TRUE} \code{\link[qdap]{qheat}} is used to
plot.  If \code{FALSE} \code{\link[stats]{heatmap}} is used.}

\item{scale}{logical.  If \code{TRUE} scales heatmap by row.  If
\code{FALSE} no scaling occurs.}

\item{\ldots}{Other arguments passed to \code{\link[qdap]{qheat}} or
\code{\link[stats]{heatmap}}.}
}
\description{
Plots a word_position object.
}

