% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/word_stats.R
\name{plot.word_stats}
\alias{plot.word_stats}
\title{Plots a word_stats object}
\usage{
\method{plot}{word_stats}(x, label = FALSE, lab.digits = NULL, ...)
}
\arguments{
\item{x}{The word_stats object}

\item{label}{logical.  If \code{TRUE} the cells of the heat map plot will be
labeled with count and proportional values.}

\item{lab.digits}{Integer values specifying the number of digits to be
printed if \code{label} is \code{TRUE}.}

\item{\ldots}{Other arguments passed to qheat.}
}
\description{
Plots a word_stats object.
}

