\docType{data}
\name{positive.words}
\alias{positive.words}
\title{Positive Words}
\format{A vector with 2006 elements}
\description{
  A dataset containing a vector of positive words.
}
\details{
  A sentence containing more negative words would be deemed
  a negative sentence, whereas a sentence containing more
  positive words would be considered positive.
}
\references{
  Hu, M., & Liu, B. (2004). Mining opinion features in
  customer reviews. National Conference on Artificial
  Intelligence.

  \url{http://www.cs.uic.edu/~liub/FBS/sentiment-analysis.html}
}
\keyword{datasets}

