% Generated by roxygen2 (4.0.1): do not edit by hand
\name{preprocessed.check_spelling_interactive}
\alias{preprocessed.check_spelling_interactive}
\title{Check Spelling}
\usage{
\method{preprocessed}{check_spelling_interactive}(x, ...)
}
\arguments{
\item{x}{The \code{\link[qdap]{check_spelling_interactive}} object.}

\item{\ldots}{ignored}
}
\description{
View check_spelling_interactive preprocessed.
}
\details{
check_spelling_interactive Method for preprocessed
}

