\name{preprocessed.pos_by}
\alias{preprocessed.pos_by}
\title{Parts of Speech}
\usage{
\method{preprocessed}{pos_by}(x, ...)
}
\arguments{
  \item{x}{The \code{\link[qdap]{pos_by}} object.}

  \item{\ldots}{ignored}
}
\description{
View pos_by preprocessed.
}
\details{
pos_by Method for preprocessed
}

