% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/question_type.R
\name{preprocessed.question_type}
\alias{preprocessed.question_type}
\title{Question Counts}
\usage{
\method{preprocessed}{question_type}(x, ...)
}
\arguments{
\item{x}{The question_type object.}

\item{\ldots}{ignored}
}
\description{
View \code{\link[qdap]{question_type}} preprocessed.
}
\details{
question_type Method for preprocessed
}

