% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qdap-package.R
\docType{data}
\name{pres_debate_raw2012}
\alias{pres_debate_raw2012}
\title{First 2012 U.S. Presidential Debate}
\format{A data frame with 94 rows and 2 variables}
\usage{
data(pres_debate_raw2012)
}
\description{
A dataset containing the raw version of the first presidential debate.
}
\details{
\itemize{
  \item person. The speaker
  \item dialogue. The words spoken
}
}
\keyword{datasets}

