% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qdap-package.R
\docType{data}
\name{pres_debates2012}
\alias{pres_debates2012}
\title{2012 U.S. Presidential Debates}
\format{A data frame with 2912 rows and 4 variables}
\usage{
data(pres_debates2012)
}
\description{
A dataset containing a cleaned version of all three presidential debates for
the 2012 election.
}
\details{
\itemize{
  \item person. The speaker
  \item tot. Turn of talk
  \item dialogue. The words spoken
  \item time. Variable indicating which of the three debates the dialogue is from
}
}
\keyword{datasets}

