% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adjacency_matrix.R
\name{print.adjacency_matrix}
\alias{print.adjacency_matrix}
\title{Prints an adjacency_matrix Object}
\usage{
\method{print}{adjacency_matrix}(x, ...)
}
\arguments{
\item{x}{The adjacency_matrix object.}

\item{\ldots}{ignored}
}
\description{
Prints an adjacency_matrix object.
}

