% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all_words.R
\name{print.all_words}
\alias{print.all_words}
\title{Prints an all_words Object}
\usage{
\method{print}{all_words}(x, ...)
}
\arguments{
\item{x}{The all_words object.}

\item{\ldots}{ignored}
}
\description{
Prints an all_words object.
}

