% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.animated_character}
\alias{print.animated_character}
\title{Prints an animated_character  Object}
\usage{
\method{print}{animated_character}(x, pause = 0, ...)
}
\arguments{
\item{x}{The animated_character  object.}

\item{pause}{The length of time to pause between plots.}

\item{title}{The title of the plot.}

\item{layout}{\pkg{igraph} \code{layout} to use.}

\item{seed}{The seed to use in plotting the graph.}

\item{legend}{The coordinates of the legend. See
\code{\link[plotrix]{color.legend}} for more information.}

\item{legend.cex}{character expansion factor. \code{NULL} and \code{NA} are
equivalent to 1.0. See \code{\link[graphics]{mtext}} for more information.}

\item{bg}{The color to be used for the background of the device region. See
\code{\link[graphics]{par}} for more information.}

\item{net.legend.color}{The text legend color for the network plot.}

\item{\ldots}{Other Arguments passed to \code{\link[igraph]{plot.igraph}}.}
}
\description{
Prints an animated_character  object.
}

