% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lexical_classification.R
\name{print.animated_lexical_classification}
\alias{print.animated_lexical_classification}
\title{Prints an animated_lexical_classification  Object}
\usage{
\method{print}{animated_lexical_classification}(x, title = NULL,
  seed = sample(1:10000, 1), layout = layout.auto, pause = 0,
  legend = c(-0.5, -1.5, 0.5, -1.45), legend.cex = 1, bg = NULL,
  net.legend.color = "black", ...)
}
\arguments{
\item{x}{The animated_lexical_classification  object.}

\item{title}{The title of the plot.}

\item{seed}{The seed to use in plotting the graph.}

\item{layout}{\pkg{igraph} \code{layout} to use.}

\item{pause}{The length of time to pause between plots.}

\item{legend}{The coordinates of the legend. See
\code{\link[plotrix]{color.legend}} for more information.}

\item{legend.cex}{character expansion factor. \code{NULL} and \code{NA} are
equivalent to 1.0. See \code{\link[graphics]{mtext}} for more information.}

\item{bg}{The color to be used for the background of the device region. See
\code{\link[graphics]{par}} for more information.}

\item{net.legend.color}{The text legend color for the network plot.}

\item{\ldots}{Other Arguments passed to \code{\link[igraph]{plot.igraph}}.}
}
\description{
Prints an animated_lexical_classification  object.
}

