% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/automated_readability_index.R
\name{print.automated_readability_index}
\alias{print.automated_readability_index}
\title{Prints an automated_readability_index Object}
\usage{
\method{print}{automated_readability_index}(x, digits = 3, ...)
}
\arguments{
\item{x}{The automated_readability_index object.}

\item{digits}{The number of digits displayed if \code{values} is \code{TRUE}.}

\item{\ldots}{ignored}
}
\description{
Prints an automated_readability_index object.
}

