% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_spelling.R
\name{print.check_spelling}
\alias{print.check_spelling}
\title{Prints a check_spelling Object}
\usage{
\method{print}{check_spelling}(x, ...)
}
\arguments{
\item{x}{The check_spelling object.}

\item{\ldots}{ignored}
}
\description{
Prints a check_spelling object.
}

