% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_spelling.R
\name{print.check_spelling_interactive}
\alias{print.check_spelling_interactive}
\title{Prints a check_spelling_interactive Object}
\usage{
\method{print}{check_spelling_interactive}(x, ...)
}
\arguments{
\item{x}{The check_spelling_interactive object.}

\item{\ldots}{ignored}
}
\description{
Prints a check_spelling_interactive object.
}

