% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cm_distance.R
\name{print.cm_distance}
\alias{print.cm_distance}
\title{Prints a cm_distance Object}
\usage{
\method{print}{cm_distance}(x, mean.digits = 0, sd.digits = 2,
  sd.mean.digits = 3, pval.digits = 3, new.order = NULL,
  na.replace = "-", diag.replace = na.replace, print = TRUE, ...)
}
\arguments{
\item{x}{The cm_distance object.}

\item{mean.digits}{The number of digits to print for the mean code distances.}

\item{sd.digits}{The number of digits to print for the standard deviations of
the code distances.}

\item{sd.mean.digits}{The number of digits to print for the standardized mean
distances.}

\item{pval.digits}{The number of digits to print for the p-values.}

\item{new.order}{An integer vector reordering the columns and rows of the
output.  Omission of a column number will result in omission from the output.}

\item{na.replace}{A character to replace \code{NA} values with.}

\item{diag.replace}{A character to replace the diagonal of the mean distance
matrix.}

\item{print}{logical.  If \code{TRUE} prints to the console.  \code{FALSE}
may be used to extract the invisibly returned output without printing to the
console.}

\item{\ldots}{ignored}
}
\description{
Prints a cm_distance object.
}

