% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.combo_syllable_sum}
\alias{print.combo_syllable_sum}
\title{Prints an combo_syllable_sum object}
\usage{
\method{print}{combo_syllable_sum}(x, ...)
}
\arguments{
\item{x}{The combo_syllable_sum object}

\item{\ldots}{ignored}
}
\description{
Prints an combo_syllable_sum object
}

