% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lexical_classification.R
\name{print.cumulative_animated_lexical_classification}
\alias{print.cumulative_animated_lexical_classification}
\title{Prints a cumulative_animated_lexical_classification Object}
\usage{
\method{print}{cumulative_animated_lexical_classification}(x, ...)
}
\arguments{
\item{x}{The cumulative_animated_lexical_classification object.}

\item{\ldots}{ignored}
}
\description{
Prints a cumulative_animated_lexical_classification  object.
}

