% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polarity.R
\name{print.cumulative_animated_polarity}
\alias{print.cumulative_animated_polarity}
\title{Prints a cumulative_animated_polarity Object}
\usage{
\method{print}{cumulative_animated_polarity}(x, ...)
}
\arguments{
\item{x}{The cumulative_animated_polarity object.}

\item{\ldots}{ignored}
}
\description{
Prints a cumulative_animated_polarity  object.
}

