% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.cumulative_formality}
\alias{print.cumulative_formality}
\title{Prints a cumulative_formality Object}
\usage{
\method{print}{cumulative_formality}(x, ...)
}
\arguments{
\item{x}{The cumulative_formality object.}

\item{\ldots}{ignored}
}
\description{
Prints a cumulative_formality  object.
}

