\name{print.discourse_map}
\alias{print.discourse_map}
\title{Prints a discourse_map Object}
\usage{
\method{print}{discourse_map}(x, edge.curved = TRUE, ...)
}
\arguments{
  \item{x}{The discourse_map object.}

  \item{edge.curved}{logical.  If \code{TRUE} edges are
  plotted with curves.}

  \item{\ldots}{Other Arguments passed to
  \code{\link[igraph]{plot.igraph}}.}
}
\description{
Prints a discourse_map object.
}

