% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/end_mark.R
\name{print.end_mark}
\alias{print.end_mark}
\title{Prints an end_mark object}
\usage{
\method{print}{end_mark}(x, ...)
}
\arguments{
\item{x}{The end_mark object}

\item{\ldots}{ignored}
}
\description{
Prints an end_mark object
}

