% Generated by roxygen2 (4.0.0): do not edit by hand
\name{print.formality}
\alias{print.formality}
\title{Prints a formality Object}
\usage{
\method{print}{formality}(x, digits, ...)
}
\arguments{
\item{x}{The formality object.}

\item{digits}{The number of digits to print.}

\item{\ldots}{ignored}
}
\description{
Prints a formality  object.
}

