\name{print.formality_preprocessed}
\alias{print.formality_preprocessed}
\title{Prints a formality_preprocessed object}
\usage{
\method{print}{formality_preprocessed}(x, ...)
}
\arguments{
  \item{x}{The formality_preprocessed object}

  \item{\ldots}{ignored}
}
\description{
Prints a formality_preprocessed object
}

