% Generated by roxygen2 (4.0.0): do not edit by hand
\name{print.formality_scores}
\alias{print.formality_scores}
\title{Prints a formality_scores object}
\usage{
\method{print}{formality_scores}(x, ...)
}
\arguments{
\item{x}{The formality_scores object}

\item{\ldots}{ignored}
}
\description{
Prints a formality_scores object
}

