% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/automated_readability_index.R
\name{print.fry}
\alias{print.fry}
\title{Prints an fry Object}
\usage{
\method{print}{fry}(x, digits = 3, auto.label, grid, div.col, plot, ...)
}
\arguments{
\item{x}{The fry object.}

\item{digits}{The number of digits displayed if \code{values} is \code{TRUE}.}

\item{auto.label}{logical.  If \code{TRUE} labels automatically added.  If
\code{FALSE} the user clicks interactively.}

\item{grid}{logical.  If \code{TRUE} a micro grid is displayed similar to
Fry's original depiction, though this makes visualizing more difficult.}

\item{div.col}{The color of the grade level division lines.}

\item{plot}{logical.  If \code{TRUE} a graph is plotted corresponding to Fry's
graphic representation.}

\item{\ldots}{ignored}
}
\description{
Prints an fry object.
}

