% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inspect_text.R
\name{print.inspect_text}
\alias{print.inspect_text}
\title{Prints an inspect_text Object}
\usage{
\method{print}{inspect_text}(x, file = "", ...)
}
\arguments{
\item{x}{The inspect_text object.}

\item{file}{A connection, or a character string naming the file to print to. If \code{""}
(the default), prints to the standard output connection, the console unless redirected
by \code{\link[base]{sink}}.}

\item{\ldots}{Other arguments passed to \code{\link[base]{strwrap}}.}
}
\description{
Prints an inspect_text object.
}

