\name{print.kullback.leibler}
\alias{print.kullback.leibler}
\title{Prints a kullback.leibler Object.}
\usage{
  \method{print}{kullback.leibler} (x, digits = 3, ...)
}
\arguments{
  \item{x}{The kullback.leibler object}

  \item{digits}{Number of decimal places to print.}

  \item{\ldots}{ignored}
}
\description{
  Prints a kullback.leibler object.
}

