% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kullback_leibler.R
\name{print.kullback_leibler}
\alias{print.kullback_leibler}
\title{Prints a kullback_leibler Object.}
\usage{
\method{print}{kullback_leibler}(x, digits = 3, ...)
}
\arguments{
\item{x}{The kullback_leibler object}

\item{digits}{Number of decimal places to print.}

\item{\ldots}{ignored}
}
\description{
Prints a kullback_leibler object.
}

