% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.lexical_classification}
\alias{print.lexical_classification}
\title{Prints an lexical_classification Object}
\usage{
\method{print}{lexical_classification}(x, ...)
}
\arguments{
\item{x}{The lexical_classification object.}

\item{\ldots}{Other arguments passed to
\code{\link[qdap]{print.lexical_classification_by}}.}
}
\description{
Prints an lexical_classification object.
}

