% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lexical_classification.R
\name{print.lexical_classification_by}
\alias{print.lexical_classification_by}
\title{Prints a lexical_classification Object}
\usage{
\method{print}{lexical_classification_by}(x, ave.digits = 1, se.digits = 2,
  trunc = 25, ...)
}
\arguments{
\item{x}{The lexical_classification_by object.}

\item{ave.digits}{The number of average lexical distribution proportion
digits to print.}

\item{se.digits}{The number of standard error of the lexical distribution
proportion digits to print.}

\item{trunc}{The width to truncate content/function word lists.}

\item{\ldots}{ignored}
}
\description{
Prints a lexical_classification_by  object.
}

