% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lexical_classification.R
\name{print.lexical_classification_preprocessed}
\alias{print.lexical_classification_preprocessed}
\title{Prints a lexical_classification_preprocessed Object}
\usage{
\method{print}{lexical_classification_preprocessed}(x, ...)
}
\arguments{
\item{x}{The lexical_classification_preprocessed object.}

\item{\ldots}{ignored}
}
\description{
Prints a lexical_classification_preprocessed object.
}

