% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lexical_classification.R
\name{print.lexical_classification_score}
\alias{print.lexical_classification_score}
\title{Prints a lexical_classification_score Object}
\usage{
\method{print}{lexical_classification_score}(x, digits = 3, ...)
}
\arguments{
\item{x}{The lexical_classification_score object.}

\item{digits}{The number of digits displayed if \code{values} is \code{TRUE}.}

\item{\ldots}{ignored}
}
\description{
Prints a lexical_classification_score object.
}

