% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/automated_readability_index.R
\name{print.linsear_write}
\alias{print.linsear_write}
\title{Prints an linsear_write Object}
\usage{
\method{print}{linsear_write}(x, digits = 3, ...)
}
\arguments{
\item{x}{The linsear_write object.}

\item{digits}{The number of digits displayed if \code{values} is \code{TRUE}.}

\item{\ldots}{ignored}
}
\description{
Prints an linsear_write object.
}

