% Generated by roxygen2 (4.0.0): do not edit by hand
\name{print.Network_polarity}
\alias{print.Network_polarity}
\title{Prints a Network_polarity  Object}
\usage{
\method{print}{Network_polarity}(x, title = NULL, title.color = "black",
  seed = sample(1:10000, 1), layout = layout.auto, legend = c(-0.5, -1.5,
  0.5, -1.45), legend.cex = 1, bg = NULL, net.legend.color = "black",
  vertex.color = "grey40", vertex.size = 20,
  vertex.label.color = "deepskyblue", vertex.label.cex = 1.1, ...)
}
\arguments{
\item{x}{The Network_polarity  object.}

\item{title}{The title of the plot.}

\item{layout}{\pkg{igraph} \code{layout} to use.}

\item{seed}{The seed to use in plotting the graph.}

\item{legend}{The coordinates of the legend. See
\code{\link[plotrix]{color.legend}} for more information.}

\item{legend.cex}{character expansion factor. \code{NULL} and \code{NA} are
equivalent to 1.0. See \code{\link[graphics]{mtext}} for more information.}

\item{bg}{The color to be used for the background of the device region. See
\code{\link[graphics]{par}} for more information.}

\item{net.legend.color}{The text legend color for the network plot.}

\item{vertex.color}{The font family to be used for vertex labels.}

\item{vertex.size}{The size of the vertex.}

\item{vertex.label.color}{The color of the labels.}

\item{vertex.label.cex}{The font size for vertex labels.}

\item{\ldots}{Other Arguments passed to \code{\link[igraph]{plot.igraph}}.}
}
\description{
Prints a Network_polarity  object.
}

