% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ngrams.R
\name{print.ngrams}
\alias{print.ngrams}
\title{Prints an ngrams object}
\usage{
\method{print}{ngrams}(x, ...)
}
\arguments{
\item{x}{The ngrams object}

\item{\ldots}{ignored}
}
\description{
Prints an ngrams object
}

