% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/object_pronoun_type.R
\name{print.object_pronoun_type}
\alias{print.object_pronoun_type}
\title{Prints a object_pronoun_type object}
\usage{
\method{print}{object_pronoun_type}(x, ...)
}
\arguments{
\item{x}{The object_pronoun_type object}

\item{\ldots}{ignored}
}
\description{
Prints a object_pronoun_type object
}

