\name{print.phrase_net}
\alias{print.phrase_net}
\title{Prints an phrase_net Object}
\usage{
\method{print}{phrase_net}(x, edge.curved = TRUE, ...)
}
\arguments{
  \item{x}{The phrase_net object.}

  \item{edge.curved}{logical.  If \code{TRUE} edges are
  plotted with curves.}

  \item{\ldots}{Other Arguments passed to
  \code{\link[igraph]{plot.igraph}}.}
}
\description{
Prints an phrase_net object.
}

