% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.polarity}
\alias{print.polarity}
\title{Prints an polarity Object}
\usage{
\method{print}{polarity}(x, digits = 3, ...)
}
\arguments{
\item{x}{The polarity object.}

\item{digits}{The number of digits displayed if \code{values} is \code{TRUE}.}

\item{\ldots}{ignored}
}
\description{
Prints an polarity object.
}

