% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polarity.R
\name{print.polarity_count}
\alias{print.polarity_count}
\title{Prints a polarity_count Object}
\usage{
\method{print}{polarity_count}(x, digits = 3, ...)
}
\arguments{
\item{x}{The polarity_count object.}

\item{digits}{The number of digits displayed.}

\item{\ldots}{ignored}
}
\description{
Prints a polarity_count object.
}

