\name{print.polarity_score}
\alias{print.polarity_score}
\title{Prints a polarity_score Object}
\usage{
\method{print}{polarity_score}(x, digits = NULL, ...)
}
\arguments{
  \item{x}{The polarity_score object.}

  \item{digits}{The number of digits displayed if
  \code{values} is \code{TRUE}.}

  \item{\ldots}{ignored}
}
\description{
Prints a polarity_score object.
}

