% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos.R
\name{print.pos}
\alias{print.pos}
\title{Prints a pos Object.}
\usage{
\method{print}{pos}(x, digits = 1, percent = NULL, zero.replace = NULL, ...)
}
\arguments{
\item{x}{The pos object}

\item{digits}{Integer values specifying the number of digits to be 
printed.}

\item{percent}{logical.  If TRUE output given as percent.  If FALSE the 
output is proportion.  If NULL uses the value from 
\code{\link[qdap]{termco}}.  Only used if \code{label} is TRUE.}

\item{zero.replace}{Value to replace 0 values with.  If NULL uses the value 
from \code{\link[qdap]{termco}}.  Only used if \code{label} is TRUE.}

\item{\ldots}{ignored}
}
\description{
Prints a pos object.
}
