% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pronoun_type.R
\name{print.pronoun_type}
\alias{print.pronoun_type}
\title{Prints a pronoun_type object}
\usage{
\method{print}{pronoun_type}(x, ...)
}
\arguments{
\item{x}{The pronoun_type object}

\item{\ldots}{ignored}
}
\description{
Prints a pronoun_type object
}

