% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/new_project.R
\name{print.qdapProj}
\alias{print.qdapProj}
\title{Prints a qdapProj Object}
\usage{
\method{print}{qdapProj}(x, ...)
}
\arguments{
\item{x}{The qdapProj object.}

\item{\ldots}{ignored}
}
\description{
Prints a qdapProj object.
}

