% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/question_type.R
\name{print.question_type}
\alias{print.question_type}
\title{Prints a question_type object}
\usage{
\method{print}{question_type}(x, ...)
}
\arguments{
\item{x}{The question_type object}

\item{\ldots}{ignored}
}
\description{
Prints a question_type object
}

