% Generated by roxygen2 (4.0.0): do not edit by hand
\name{print.question_type_preprocessed}
\alias{print.question_type_preprocessed}
\title{Prints a question_type_preprocessed object}
\usage{
\method{print}{question_type_preprocessed}(x, ...)
}
\arguments{
\item{x}{The question_type_preprocessed object}

\item{\ldots}{ignored}
}
\description{
Prints a question_type_preprocessed object
}

