\name{print.readability_count}
\alias{print.readability_count}
\title{Prints a readability_count Object}
\usage{
\method{print}{readability_count}(x, digits = 3, ...)
}
\arguments{
  \item{x}{The readability_count object.}

  \item{digits}{The number of digits displayed.}

  \item{\ldots}{ignored}
}
\description{
Prints a readability_count object.
}

