% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/automated_readability_index.R
\name{print.readability_score}
\alias{print.readability_score}
\title{Prints a readability_score Object}
\usage{
\method{print}{readability_score}(x, digits = 3, ...)
}
\arguments{
\item{x}{The readability_score object.}

\item{digits}{The number of digits displayed if \code{values} is \code{TRUE}.}

\item{\ldots}{ignored}
}
\description{
Prints a readability_score object.
}

