% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sentSplit.R
\name{print.sent_split}
\alias{print.sent_split}
\title{Prints a sent_split object}
\usage{
\method{print}{sent_split}(x, ...)
}
\arguments{
\item{x}{The sent_split object}

\item{\ldots}{ignored}
}
\description{
Prints a sent_split object
}

