% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multigsub.R
\name{print.sub_holder}
\alias{print.sub_holder}
\title{Prints a sub_holder object}
\usage{
\method{print}{sub_holder}(x, ...)
}
\arguments{
\item{x}{The sub_holder object}

\item{\ldots}{ignored}
}
\description{
Prints a sub_holder object
}

