% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.subject_pronoun_type}
\alias{print.subject_pronoun_type}
\title{Prints a subject_pronoun_type object}
\usage{
\method{print}{subject_pronoun_type}(x, ...)
}
\arguments{
\item{x}{The subject_pronoun_type object}

\item{\ldots}{ignored}
}
\description{
Prints a subject_pronoun_type object
}

