% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gantt.R
\name{print.sums_gantt}
\alias{print.sums_gantt}
\title{Prints a sums_gantt object}
\usage{
\method{print}{sums_gantt}(x, ...)
}
\arguments{
\item{x}{The sums_gantt object}

\item{\ldots}{ignored}
}
\description{
Prints a sums_gantt object.
}

