% Generated by roxygen2 (4.0.0): do not edit by hand
\name{print.syllable_sum}
\alias{print.syllable_sum}
\title{Prints an syllable_sum object}
\usage{
\method{print}{syllable_sum}(x, ...)
}
\arguments{
\item{x}{The syllable_sum object}

\item{\ldots}{ignored}
}
\description{
Prints an syllable_sum object
}

