% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/table_methods.R
\name{print.table_count}
\alias{print.table_count}
\title{Prints a table_count object}
\usage{
\method{print}{table_count}(x, ...)
}
\arguments{
\item{x}{The table_count object}

\item{\ldots}{ignored}
}
\description{
Prints a table_count object
}

