% Generated by roxygen2 (4.0.0): do not edit by hand
\name{print.table_proportion}
\alias{print.table_proportion}
\title{Prints a table_proportion object}
\usage{
\method{print}{table_proportion}(x, ...)
}
\arguments{
\item{x}{The table_proportion object}

\item{\ldots}{ignored}
}
\description{
Prints a table_proportion object
}

