% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/termco.R
\name{print.termco}
\alias{print.termco}
\title{Prints a termco object.}
\usage{
\method{print}{termco}(x, digits = NULL, percent = NULL,
  zero.replace = NULL, ...)
}
\arguments{
\item{x}{The termco object}

\item{digits}{Integer values specifying the number of digits to be
printed.}

\item{percent}{logical.  If TRUE output given as percent.  If FALSE the
output is proportion.  If NULL uses the value from
\code{\link[qdap]{termco}}.  Only used if \code{label} is TRUE.}

\item{zero.replace}{Value to replace 0 values with.  If NULL uses the value
from \code{\link[qdap]{termco}}.  Only used if \code{label} is TRUE.}

\item{\ldots}{ignored}
}
\description{
Prints a termco object.
}

