% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/htruncdf.R
\name{print.trunc}
\alias{print.trunc}
\title{Prints a trunc object}
\usage{
\method{print}{trunc}(x, ...)
}
\arguments{
\item{x}{The trunc object}

\item{\ldots}{ignored}
}
\description{
Prints a trunc object
}

