% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/type_token_ratio.R
\name{print.type_token_ratio}
\alias{print.type_token_ratio}
\title{Prints a type_token_ratio Object}
\usage{
\method{print}{type_token_ratio}(x, digits = 3, ...)
}
\arguments{
\item{x}{The type_token_ratio object.}

\item{digits}{The number of type-token ratio digits to print.}

\item{\ldots}{ignored}
}
\description{
Prints a type_token_ratio  object.
}

