\name{print.v.outer}
\alias{print.v.outer}
\title{Prints a v.outer Object.}
\usage{
  \method{print}{v.outer} (x, digits = 3, ...)
}
\arguments{
  \item{x}{The v.outer object}

  \item{digits}{Number of decimal places to print.}

  \item{\ldots}{ignored}
}
\description{
  Prints a v.outer object.
}

