\name{print.wfm}
\alias{print.wfm}
\title{Prints a wfm Object}
\usage{
\method{print}{wfm}(x, digits = 3, ...)
}
\arguments{
  \item{x}{The wfm object.}

  \item{digits}{The number of digits displayed if
  \code{values} is \code{TRUE}.}

  \item{\ldots}{ignored}
}
\description{
Prints a wfm object.
}

