% Generated by roxygen2 (4.0.0): do not edit by hand
\name{print.wfm_summary}
\alias{print.wfm_summary}
\title{Prints a wfm_summary Object}
\usage{
\method{print}{wfm_summary}(x, ...)
}
\arguments{
\item{x}{The wfm_summary object.}

\item{\ldots}{ignored}
}
\description{
Prints a wfm_summary object.
}

