% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.which_misspelled}
\alias{print.which_misspelled}
\title{Prints a which_misspelled Object}
\usage{
\method{print}{which_misspelled}(x, ...)
}
\arguments{
\item{x}{The which_misspelled object.}

\item{\ldots}{ignored}
}
\description{
Prints a which_misspelled object.
}

