% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.word_associate}
\alias{print.word_associate}
\title{Prints a word_associate object}
\usage{
\method{print}{word_associate}(x, ...)
}
\arguments{
\item{x}{The word_associate object}

\item{\ldots}{ignored}
}
\description{
Prints a word_associate object.
}

