% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/word_cor.R
\name{print.word_cor}
\alias{print.word_cor}
\title{Prints a word_cor object}
\usage{
\method{print}{word_cor}(x, digits = 3, ...)
}
\arguments{
\item{x}{The word_cor object}

\item{digits}{The number of digits to print}

\item{\ldots}{ignored}
}
\description{
Prints a word_cor object
}

