% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.word_length}
\alias{print.word_length}
\title{Prints a word_length object}
\usage{
\method{print}{word_length}(x, ...)
}
\arguments{
\item{x}{The word_length object}

\item{\ldots}{ignored}
}
\description{
Prints a word_length object
}

