% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/word_list.R
\name{print.word_list}
\alias{print.word_list}
\title{Prints a word_list Object}
\usage{
\method{print}{word_list}(x, ...)
}
\arguments{
\item{x}{The word_list object}

\item{\ldots}{ignored}
}
\description{
Prints a word_list object.
}

