\name{print.word_proximity}
\alias{print.word_proximity}
\title{Prints a word_proximity object}
\usage{
\method{print}{word_proximity}(x, digits = 3, ...)
}
\arguments{
  \item{x}{The word_proximity object}

  \item{digits}{The number of duguts to print}

  \item{\ldots}{ignored}
}
\description{
Prints a word_proximity object
}

