% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/word_stats.R
\name{print.word_stats}
\alias{print.word_stats}
\title{Prints a word_stats object}
\usage{
\method{print}{word_stats}(x, digits = NULL, ...)
}
\arguments{
\item{x}{The word_stats object}

\item{digits}{Integer; number of decimal places to round in the display of
the output.}

\item{\ldots}{ignored}
}
\description{
Prints a word_stats object.
}

