% Generated by roxygen2 (4.0.1): do not edit by hand
\name{proportions.character_table}
\alias{proportions.character_table}
\title{Term Counts}
\usage{
\method{proportions}{character_table}(x, ...)
}
\arguments{
\item{x}{The character_table object.}

\item{\ldots}{ignored}
}
\description{
View \code{\link[qdap]{character_table}} proportions.
}
\details{
character_table Method for proportions
}

