% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formality.R
\name{proportions.formality}
\alias{proportions.formality}
\title{Formality}
\usage{
\method{proportions}{formality}(x, ...)
}
\arguments{
\item{x}{The formality object.}

\item{\ldots}{ignored}
}
\description{
View \code{\link[qdap]{formality}} proportions.
}
\details{
formality Method for proportions
}
