% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos.R
\name{proportions.pos}
\alias{proportions.pos}
\title{Parts of Speech}
\usage{
\method{proportions}{pos}(x, ...)
}
\arguments{
\item{x}{The pos object.}

\item{\ldots}{ignored}
}
\description{
View \code{\link[qdap]{pos}} proportions.
}
\details{
pos Method for proportions
}
