% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/termco.R
\name{proportions.termco}
\alias{proportions.termco}
\title{Term Counts}
\usage{
\method{proportions}{termco}(x, ...)
}
\arguments{
\item{x}{The termco object.}

\item{\ldots}{ignored}
}
\description{
View \code{\link[qdap]{termco}} proportions.
}
\details{
termco Method for proportions
}

