% Generated by roxygen2 (4.0.2): do not edit by hand
\name{proportions.word_length}
\alias{proportions.word_length}
\title{Word Length Counts}
\usage{
\method{proportions}{word_length}(x, ...)
}
\arguments{
\item{x}{The word_length object.}

\item{\ldots}{ignored}
}
\description{
View \code{\link[qdap]{word_length}} proportions.
}
\details{
word_length Method for proportions
}

