% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/word_position.R
\name{proportions.word_position}
\alias{proportions.word_position}
\title{Word Position}
\usage{
\method{proportions}{word_position}(x, ...)
}
\arguments{
\item{x}{The word_position object.}

\item{\ldots}{ignored}
}
\description{
View \code{\link[qdap]{word_position}} proportions.
}
\details{
word_position Method for proportions
}

