\name{qdap-package}
\alias{qdap-package}
\alias{qdap}
\docType{package}
\title{
Aids in Quantitative Discourse Analysis
}
\description{
qdap (Quantitative Discourse Analysis Package) is an R package designed to assist in quantitative discourse analysis. The package stands as a bridge between qualitative transcripts of dialogue and statistical analysis and visualization.
}
\details{
\tabular{ll}{
Package: \tab qdap\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-07-01\cr
License: \tab GPL-2\cr
}
}
\author{
Tyler Rinker

Maintainer: Tyler Rinker <tyler.rinker@gmail.com>
}

