% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qdap-package.R
\docType{package}
\name{qdap}
\alias{package-qdap}
\alias{qdap}
\alias{qdap-package}
\title{qdap: Quantitative Discourse Analysis Package}
\description{
This package automates many of the tasks associated with quantitative
discourse analysis of transcripts containing discourse.  The package
provides parsing tools for preparing transcript data, coding tools and
analysis tools for richer understanding of the data.  Many functions
allow the user to aggregate data by any number of grouping variables,
providing analysis and seamless integration with other R packages which
enable higher level analysis and visualization of text.  This empowers
the researcher with more flexible, efficient and targeted methods and tools.
}

