% Generated by roxygen2 (4.0.1): do not edit by hand
\name{qdap_df}
\alias{Text}
\alias{Text<-}
\alias{qdap_df}
\title{Create qdap Specific Data Structure}
\usage{
qdap_df(dataframe, text.var)

Text(object)

Text(object) <- value
}
\arguments{
\item{dataframe}{A \code{\link[base]{data.frame}} with a text variable.
Generally, \code{\link[qdap]{sentSplit}} should be run first
(\code{\link[qdap]{sentSplit}}) actually produces a
\code{\link[base]{data.frame}} that is of the class \code{"qdap_df"}).}

\item{text.var}{The name of the \code{text.var} column.}

\item{object}{A \code{\link[base]{data.frame}} of the class \code{"qdap_df"}.}

\item{value}{A character string of the updated \code{text.var} column.}
}
\value{
Returns a \code{\link[base]{data.frame}} of the class \code{"qdap_df"}.
}
\description{
Creating this \pkg{qdap} specific data structure enables short hand with
subsequent \pkg{qdap} function calls that utilize the \code{text.var}
argument.  Combined with the \code{\link[qdap]{\%&\%}} operator, the user n
need not specify a data set or the \code{text.var} argument (as many
\pkg{qdap} functions contain a \code{text.var} argument).

Change text.var column of a qdap_df object.
}
\examples{
\dontrun{
dat <- qdap_df(DATA, state)
dat \%&\% trans_cloud(grouping.var=person)
dat \%&\% trans_cloud(grouping.var=person, text.var=stemmer(DATA$state))
dat \%&\% termco(grouping.var=person, match.list=list("fun", "computer"))
class(dat)

dat2 <- sentSplit(DATA, "state", stem.col = TRUE)
class(dat2)
dat2 \%&\% trans_cloud()
Text(dat2)
## change the `text.var` column
Text(dat2) <- "stem.text"
dat2 \%&\% trans_cloud()

dat <- sentSplit(DATA, "state")
dat \%&\% trans_cloud(grouping.var=person)
dat \%&\% termco(person, match.list=list("fun", "computer"))
dat \%&\% trans_venn(person)
dat \%&\% polarity(person)
dat \%&\% formality(person)
dat \%&\% automated_readability_index(person)
dat \%&\% Dissimilarity(person)
dat \%&\% gradient_cloud(sex)
dat \%&\% dispersion_plot(c("fun", "computer"))
dat \%&\% discourse_map(list(sex, adult))
dat \%&\% gantt_plot(person)
dat \%&\% word_list(adult)
dat \%&\% end_mark_by(person)
dat \%&\% end_mark()
dat \%&\% word_stats(person)
dat \%&\% wfm(person)
dat \%&\% word_cor(person, "i")
dat \%&\% sentCombine(person)
dat \%&\% question_type(person)

## combine with magrittr/dplyr
library(magrittr)
dat \%&\% wfm(person) \%>\% plot()
dat \%&\% polarity(person) \%>\% scores()
dat \%&\% polarity(person) \%>\% counts()
dat \%&\% polarity(person) \%>\% scores()
}
}
\references{
Inspired by \pkg{dplyr}'s \code{\link[dplyr]{tbl_df}} structure.
}
\seealso{
\code{\link[qdap]{\%&\%}},
\code{\link[qdap]{sentSplit}}
}
\keyword{Text}
\keyword{data}
\keyword{structure}

