% Generated by roxygen2 (4.0.0): do not edit by hand
\name{qtheme}
\alias{qtheme}
\alias{theme_nightheat}
\title{Add themes to a Network object.}
\usage{
qtheme(x = "generic", title, title.color, layout, legend, legend.cex,
  legend.text.color, bg, vertex.color, vertex.size, vertex.label.color,
  vertex.label.cex, edge.label.color, edge.label.cex)

theme_nightheat(x = pars[["x"]], title = pars[["title"]],
  title.color = pars[["title.color"]], layout = pars[["layout"]],
  legend = pars[["legend"]], legend.cex = pars[["legend.cex"]],
  bg = pars[["bg"]], legend.text.color = pars[["legend.text.color"]],
  vertex.color = pars[["vertex.color"]],
  vertex.size = pars[["vertex.size"]],
  vertex.label.color = pars[["vertex.label.color"]],
  vertex.label.cex = pars[["vertex.label.cex"]],
  edge.label.color = pars[["edge.label.color"]],
  edge.label.cex = pars[["edge.label.cex"]], ...)
}
\arguments{
\item{x}{The name of the qtheme.}

\item{title.color}{The color of the title.}

\item{layout}{\pkg{igraph} \code{layout} to use.}

\item{legend}{The coordinates of the legend. See
\code{\link[plotrix]{color.legend}} for more information.}

\item{legend.cex}{character expansion factor. \code{NULL} and \code{NA} are
equivalent to 1.0. See \code{\link[graphics]{mtext}} for more information.}

\item{legend.text.color}{The text legend text color.}

\item{bg}{The color to be used for the background of the device region. See
\code{\link[graphics]{par}} for more information.}

\item{vertex.color}{The font family to be used for vertex labels.}

\item{vertex.size}{The size of the vertex.}

\item{vertex.label.color}{The color of the labels.}

\item{vertex.label.cex}{The font size for vertex labels.}

\item{edge.label.color}{The color for the edge labels.  Use \code{NA} to
remove.}

\item{edge.label.cex}{The font size of the edge labels.}

\item{title}{The title of the plot.  \code{NULL} eliminates title.  \code{NA}
uses title attribute of the Network object.}

\item{\ldots}{Additional arguments supplied to \code{qtheme}.}
}
\description{
This function builds generic themes to add a theme to a Network object rather
than individual \code{print} arguments.

This theme allows you to add a night heat theme to a Network object rather
than individual \code{print} arguments.
}

