\docType{data}
\name{raj.act.1}
\alias{raj.act.1}
\title{Romeo and Juliet: Act 1}
\format{A data frame with 235 rows and 2 variables}
\description{
  A dataset containing Romeo and Juliet: Act 1.
}
\details{
  \itemize{ \item person. Character in the play \item
  dialogue. The spoken dialogue }
}
\references{
  \url{http://shakespeare.mit.edu/romeo_juliet/full.html}
}
\keyword{datasets}

