\docType{data}
\name{raj.act.5}
\alias{raj.act.5}
\title{Romeo and Juliet: Act 5}
\format{A data frame with 88 rows and 2 variables}
\description{
  A dataset containing Romeo and Juliet: Act 5.
}
\details{
  \itemize{ \item person. Character in the play \item
  dialogue. The spoken dialogue }
}
\references{
  \url{http://shakespeare.mit.edu/romeo_juliet/full.html}
}
\keyword{datasets}

