\docType{data}
\name{raj}
\alias{raj}
\title{Romeo and Juliet (Unchanged & Complete)}
\format{A data frame with 840 rows and 3 variables}
\description{
  A dataset containing the original transcript from Romeo
  and Juliet as it was scraped from:
  \url{http://shakespeare.mit.edu/romeo_juliet/full.html}.
}
\details{
  \itemize{ \item person. Character in the play \item
  dialogue. The spoken dialogue \item act. The act (akin to
  repeated measures) }
}
\references{
  \url{http://shakespeare.mit.edu/romeo_juliet/full.html}
}
\keyword{datasets}

