\name{rajPOS}
\alias{rajPOS}
\docType{data}
\title{
Romeo and Juliet parts of speech
}
\description{
Romeo and Juliet broken into parts of speech
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}

