% Generated by roxygen2 (4.0.1): do not edit by hand
\name{random_sent}
\alias{random_sent}
\title{Generate Random Sentence Samples}
\usage{
random_sent(n = 10, len = 14, range = len - 1,
  dictionary = qdapDictionaries::Top200Words, endmark.fun = function()
  sample(c(".", "!", "|", "?"), 1, prob = c(0.85, 0.05, 0.05, 0.05)))
}
\arguments{
\item{n}{Number of sentences to create.}

\item{len}{Average length of sentences (in words).}

\item{range}{Range around \code{len} that number of words may vary.  This may
be a recycled single integer vector or an integer vector of length 2.}

\item{dictionary}{A dictionary of words to sample from.}

\item{endmark.fun}{A function to create random end marks.}
}
\value{
Returns a random vector of sentence strings.
}
\description{
Generates a random sample of sentences (sentences are sampled at the word
level and there for are likely nonsensical).
}
\examples{
\dontrun{
random_sent()
random_sent(200, 10)

dict <- sort(unique(bag_o_words(pres_debates2012[["dialogue"]])))
random_sent(dictionary=dict)
}
}
\keyword{random}
\keyword{sample}
\keyword{sentence}

