\name{raw}
\alias{raw}
\title{Generic Raw Method}
\usage{
raw(x, ...)
}
\arguments{
  \item{x}{A qdap object (list) with a raw dataframe (e.g.,
  \code{\link[qdap]{pos_by}}).}

  \item{\ldots}{Arguments passed to proportions method of
  other classes.}
}
\value{
Returns a data.frame or list of raw processed data.
}
\description{
Access the raw dataframes from select qdap outputs.
}
\seealso{
\code{\link[qdap]{scores}}

\code{\link[qdap]{counts}}

\code{\link[qdap]{proportions}}
}

