% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{raw.time.span}
\alias{raw.time.span}
\title{Minimal Raw Time Span Data Set}
\format{A list with 3 elements}
\usage{
data(raw.time.span)
}
\description{
A dataset containing a list of named vectors of time spans.
}
\keyword{datasets}

