% Generated by roxygen2 (4.0.1): do not edit by hand
\name{replace_ordinal}
\alias{replace_ordinal}
\title{Replace Mixed Ordinal Numbers With Text Representation}
\usage{
replace_ordinal(text.var, num.paste = TRUE, remove = FALSE)
}
\arguments{
\item{text.var}{The text variable.}

\item{num.paste}{logical.  If \code{TRUE} a the elements of larger numbers are
separated with spaces.  If \code{FALSE} the elements will be joined without
spaces.}

\item{remove}{logical.  If \code{TRUE} ordinal numbers are removed from the text.}
}
\description{
Replaces mixed text/numeric represented ordinal numbers with words (e.g.,
"1st" becomes "first").
}
\note{
Currently only implemented for ordinam values: 1-25, 30, 40, 50, 60, 70, 80, 90, & 100
}
\examples{
\dontrun{
x <- "I like the 1st one not the 22nd one."
replace_ordinal(x)
}
}
\seealso{
\code{\link[qdap]{bracketX}},
\code{\link[qdap]{qprep}},
\code{\link[qdap]{replace_abbreviation}},
\code{\link[qdap]{replace_contraction}},
\code{\link[qdap]{replace_symbol}},
\code{\link[qdap]{replace_number}}
\code{\link[english]{english}}
}
\keyword{ordinal-to-word}

