% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rm_phone}
\alias{rm_phone}
\title{Remove/Replace/Extract Phone Numbers}
\usage{
rm_phone(text.var, trim = TRUE, clean = TRUE,
  pattern = "(?:(?:\\\\+?1\\\\s*(?:[.-]\\\\s*)?)?(?:\\\\(\\\\s*([2-9]1[02-9]|[2-9][02-8]1|[2-9][02-8][02-9])\\\\s*\\\\)|([2-9]1[02-9]|[2-9][02-8]1|[2-9][02-8][02-9]))\\\\s*(?:[.-]\\\\s*)?)?([2-9]1[02-9]|[2-9][02-9]1|[2-9][02-9]{2})\\\\s*(?:[.-]\\\\s*)?([0-9]{4})(?:\\\\s*(?:#|x\\\\.?|ext\\\\.?|extension)\\\\s*(\\\\d+))?",
  replacement = "", extract = FALSE, ...)
}
\arguments{
\item{text.var}{The text variable.}

\item{trim}{logical.  If \code{TRUE} removes leading and trailing white
spaces.}

\item{clean}{trim logical.  If \code{TRUE} extra white spaces and escaped
character will be removed.}

\item{pattern}{A character string containing a regular expression (or
character string for \code{fixed = TRUE}) to be matched in the given
character vector.}

\item{replacement}{Replacement for matched \code{pattern}.}

\item{extract}{logical.  If \code{TRUE} the times are extracted into a
list of vectors.}

\item{\dots}{Other arguments passed to \code{\link[base]{gsub}}.}
}
\value{
Returns a character string with phone numbers removed.
}
\description{
Remove/replace/extract phone numbers from a string.
}
\examples{
\dontrun{
x <- c(" Mr. Bean bought 2 tickets 2-613-213-4567 or 5555555555 call either one",
  "43 Butter Rd, Brossard QC K0A 3P0 - 613 213 4567",
  "Please contact Mr. Bean (613)2134567",
  "1.575.555.5555 is his #1 number",
  "7164347566",
  "I like 1234567 dogs"
)

rm_phone(x)
rm_phone(x, extract=TRUE)
}
}
\author{
\href{http://stackoverflow.com/}{stackoverflow's} Marius and Tyler Rinker <tyler.rinker@gmail.com>.
}
\references{
The time regular expression was taken from:
\url{http://stackoverflow.com/a/21008254/1000343}
}
\seealso{
\code{\link[base]{gsub}}
}
\keyword{time}

