% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rm_time}
\alias{rm_time}
\title{Remove/Replace/Extract Time}
\usage{
rm_time(text.var, trim = TRUE, clean = TRUE,
  pattern = "\\\\d{0,2}:\\\\d{2}(?:[:.]\\\\d+)?", replacement = "",
  extract = FALSE, ...)
}
\arguments{
\item{text.var}{The text variable.}

\item{trim}{logical.  If \code{TRUE} removes leading and trailing white
spaces.}

\item{clean}{trim logical.  If \code{TRUE} extra white spaces and escaped
character will be removed.}

\item{pattern}{A character string containing a regular expression (or
character string for \code{fixed = TRUE}) to be matched in the given
character vector.}

\item{replacement}{Replacement for matched \code{pattern}.}

\item{extract}{logical.  If \code{TRUE} the times are extracted into a
list of vectors.}

\item{\dots}{Other arguments passed to \code{\link[base]{gsub}}.}
}
\value{
Returns a character string with time removed.
}
\description{
Remove/replace/extract time from a string.
}
\examples{
\dontrun{
x <-  c("R uses 1:5 for 1, 2, 3, 4, 5.",
    "At 3:00 we'll meet up and leave by 4:30:20",
    "We'll meet at 6:33.", "He ran it in :22.34")

rm_time(x)
rm_time(x, extract=TRUE)
}
}
\author{
\href{http://stackoverflow.com/}{stackoverflow's} hwnd and Tyler Rinker <tyler.rinker@gmail.com>.
}
\references{
The time regular expression was taken from:
\url{http://stackoverflow.com/a/25111133/1000343}
}
\seealso{
\code{\link[base]{gsub}}
}
\keyword{time}

