% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rm_zip}
\alias{rm_zip}
\title{Remove/Replace/Extract Zip Codes}
\usage{
rm_zip(text.var, trim = TRUE, clean = TRUE,
  pattern = "(?<!\\\\d)\\\\d{5}(?:[ -]\\\\d{4})?\\\\b", replacement = "",
  extract = FALSE, ...)
}
\arguments{
\item{text.var}{The text variable.}

\item{trim}{logical.  If \code{TRUE} removes leading and trailing white
spaces.}

\item{clean}{trim logical.  If \code{TRUE} extra white spaces and escaped
character will be removed.}

\item{pattern}{A character string containing a regular expression (or
character string for \code{fixed = TRUE}) to be matched in the given
character vector.}

\item{replacement}{Replacement for matched \code{pattern}.}

\item{extract}{logical.  If \code{TRUE} the times are extracted into a
list of vectors.}

\item{\dots}{Other arguments passed to \code{\link[base]{gsub}}.}
}
\value{
Returns a character string with zip codes removed.
}
\description{
Remove/replace/extract zip codes from a string.
}
\examples{
\dontrun{
x <- c("Mr. Bean bought 2 tickets 2-613-213-4567",
  "43 Butter Rd, Brossard QC K0A 3P0 - 613 213 4567",
  "Rat Race, XX, 12345",
  "Ignore phone numbers(613)2134567",
  "Grab zips with dashes 12345-6789 or no space before12345-6789",
  "Grab zips with spaces 12345 6789 or no space before12345 6789",
  "I like 1234567 dogs"
)

rm_zip(x)
rm_zip(x, extract=TRUE)
}
}
\author{
\href{http://stackoverflow.com/}{stackoverflow's} hwnd and Tyler Rinker <tyler.rinker@gmail.com>.
}
\references{
The time regular expression was taken from:
\url{http://stackoverflow.com/a/25223890/1000343}
}
\seealso{
\code{\link[base]{gsub}}
}
\keyword{time}

