% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qdap-package.R
\docType{data}
\name{sample.time.span}
\alias{sample.time.span}
\title{Minimal Time Span Data Set}
\format{A data frame with 9 rows and 6 variables}
\usage{
data(sample.time.span)
}
\description{
A fictitious dataset containing time spans for codes A and B.
}
\details{
\itemize{
  \item code. The qualitative code.
  \item start. The integer start time.
  \item end. The integer end time.
  \item Start. The chron start time.
  \item End. The chron end time.
  \item variable. An arbitrary single time repeated measures variable (ignore).
}
}
\keyword{datasets}

