\name{scores.character_table}
\alias{scores.character_table}
\title{Term Counts}
\usage{
\method{scores}{character_table}(x, ...)
}
\arguments{
  \item{x}{The \code{\link[qdap]{character_table}} object.}

  \item{\ldots}{ignored}
}
\description{
View character_table scores.
}
\details{
character_table Method for scores
}

