\name{scores.flesch_kincaid}
\alias{scores.flesch_kincaid}
\title{Readability Measures}
\usage{
\method{scores}{flesch_kincaid}(x, ...)
}
\arguments{
  \item{x}{The flesch_kincaid object.}

  \item{\ldots}{ignored}
}
\description{
\code{scores.flesch_kincaid} - View scores from
\code{\link[qdap]{flesch_kincaid}}.
}
\details{
flesch_kincaid Method for scores
}

