% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scores.object_pronoun_type}
\alias{scores.object_pronoun_type}
\title{Question Counts}
\usage{
\method{scores}{object_pronoun_type}(x, ...)
}
\arguments{
\item{x}{The \code{\link[qdap]{object_pronoun_type}} object.}

\item{\ldots}{ignored}
}
\description{
View object_pronoun_type scores.
}
\details{
object_pronoun_type Method for scores
}

