% Generated by roxygen2 (4.0.0): do not edit by hand
\name{scores.polarity}
\alias{scores.polarity}
\title{Polarity}
\usage{
\method{scores}{polarity}(x, ...)
}
\arguments{
\item{x}{The polarity object.}

\item{\ldots}{ignored}
}
\description{
\code{scores.polarity} - View scores from \code{\link[qdap]{polarity}}.
}
\details{
polarity Method for scores
}

