% Generated by roxygen2 (4.0.1): do not edit by hand
\name{scores.pos_by}
\alias{scores.pos_by}
\title{Parts of Speech}
\usage{
\method{scores}{pos_by}(x, ...)
}
\arguments{
\item{x}{The \code{\link[qdap]{pos_by}} object.}

\item{\ldots}{ignored}
}
\description{
View pos_by scores.
}
\details{
pos_by Method for scores
}

