% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pronoun_type.R
\name{scores.pronoun_type}
\alias{scores.pronoun_type}
\title{Question Counts}
\usage{
\method{scores}{pronoun_type}(x, ...)
}
\arguments{
\item{x}{The \code{\link[qdap]{pronoun_type}} object.}

\item{\ldots}{ignored}
}
\description{
View pronoun_type scores.
}
\details{
pronoun_type Method for scores
}

