% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scores.R
\name{scores}
\alias{scores}
\title{Generic Scores Method}
\usage{
scores(x, ...)
}
\arguments{
\item{x}{A qdap object (list) with a dataframe of scores (e.g.,
\code{\link[qdap]{fry}}, \code{\link[qdap]{formality}}).}

\item{\ldots}{Arguments passed to scores method of other classes.}
}
\value{
Returns a data.frame of scores.
}
\description{
Access the scores dataframes from select qdap outputs.
}
\seealso{
\code{\link[qdap]{counts}}

\code{\link[qdap]{proportions}}

\code{\link[qdap]{preprocessed}}
}

