% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated_readability_index.R
\name{scores.SMOG}
\alias{scores.SMOG}
\title{Readability Measures}
\usage{
\method{scores}{SMOG}(x, ...)
}
\arguments{
\item{x}{The SMOG object.}

\item{\ldots}{ignored}
}
\description{
\code{scores.SMOG} - View scores from \code{\link[qdap]{SMOG}}.
}
\details{
SMOG Method for scores
}
