% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subject_pronoun_type.R
\name{scores.subject_pronoun_type}
\alias{scores.subject_pronoun_type}
\title{Question Counts}
\usage{
\method{scores}{subject_pronoun_type}(x, ...)
}
\arguments{
\item{x}{The \code{\link[qdap]{subject_pronoun_type}} object.}

\item{\ldots}{ignored}
}
\description{
View subject_pronoun_type scores.
}
\details{
subject_pronoun_type Method for scores
}

