% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/termco.R
\name{scores.termco}
\alias{scores.termco}
\title{Term Counts}
\usage{
\method{scores}{termco}(x, ...)
}
\arguments{
\item{x}{The \code{\link[qdap]{termco}} object.}

\item{\ldots}{ignored}
}
\description{
View termco scores.
}
\details{
termco Method for scores
}

