% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/word_position.R
\name{scores.word_position}
\alias{scores.word_position}
\title{Word Position}
\usage{
\method{scores}{word_position}(x, ...)
}
\arguments{
\item{x}{The \code{\link[qdap]{word_position}} object.}

\item{\ldots}{ignored}
}
\description{
View word_position scores.
}
\details{
word_position Method for scores
}

