% Generated by roxygen2 (4.0.2): do not edit by hand
\name{t.DocumentTermMatrix}
\alias{t.DocumentTermMatrix}
\title{Transposes a DocumentTermMatrix object}
\usage{
\method{t}{DocumentTermMatrix}(x, ...)
}
\arguments{
\item{x}{The DocumentTermMatrix object}

\item{\ldots}{ignored}
}
\description{
Transposes a DocumentTermMatrix object
}

