% Generated by roxygen2 (4.0.0): do not edit by hand
\name{t.TermDocumentMatrix}
\alias{t.TermDocumentMatrix}
\title{Transposes a TermDocumentMatrix object}
\usage{
\method{t}{TermDocumentMatrix}(x, ...)
}
\arguments{
\item{x}{The TermDocumentMatrix object}

\item{\ldots}{ignored}
}
\description{
Transposes a TermDocumentMatrix object
}

