\name{termco.c}
\alias{termco.c}
\title{Combine Columns from a termco_ Object}
\usage{
  termco.c(termco.d.object, combined.columns, new.name,
    short.term = FALSE, zero.replace = NULL,
    lazy.term = TRUE, elim.old = TRUE, output = "percent")
}
\arguments{
  \item{termco.d.object}{An object generated by either
  \code{termco_a} or \code{termco_d}.}

  \item{combined.columns}{The names/indexes of the columns
  to be combined.}

  \item{new.name}{A character vector of length one to name
  the new combined column.}

  \item{short.term}{logical.  If TRUE column names are
  trimmed versions of the match list, other wise the terms
  are wrapped with 'term(phrase)'}

  \item{zero.replace}{Value to replace zeros with.}

  \item{lazy.term}{logical.  If TRUE is approximate
  amtching (generally for internal use).}

  \item{elim.old}{logical.  If TRUE eliminates the columns
  that are combined together by the named match.list.}

  \item{output}{Type of proportion output; either
  \code{"proportion"} (decimal format) or \code{"percent"}.
  Default is \code{"percent"}.}
}
\value{
  Returns a return a list, of class "termco.c", of data
  frames and information regarding word counts:
  \item{raw}{raw word counts by grouping variable}
  \item{prop}{proportional word counts by grouping
  variable; proportional to each individual's word use}
  \item{rnp}{a character combination data frame of raw and
  proportional} \item{zero_replace}{value to replace zeros
  with; mostly internal use} \item{output}{character value
  for outpur type (either" "proportion" or "percent";
  mostly internal use} \item{digits}{integer value od
  number of digits to display; mostly internal use}
}
\description{
  Combines the columns of a termco_ object.  Generally
  intended for internal use but documented for
  completeness.
}
\seealso{
  \code{\link[qdap]{termco.a}}
}

