% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Trim.R
\name{Trim}
\alias{Trim}
\title{Remove Leading/Trailing White Space}
\usage{
Trim(x)
}
\arguments{
\item{x}{The text variable.}
}
\value{
Returns a vector with the leading/trailing white spaces removed.
}
\description{
Remove leading/trailing white space.
}
\examples{
\dontrun{
(x <- c("  talkstats.com ", "   really? ", " yeah"))
Trim(x)
}
}

