% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unique_by.R
\name{unique_by}
\alias{unique_by}
\title{Find Unique Words by Grouping Variable}
\usage{
unique_by(text.var, grouping.var)
}
\arguments{
\item{text.var}{The text variable}

\item{grouping.var}{The grouping variables.  Default \code{NULL} generates
one word list for all text.  Also takes a single grouping variable or a list
of 1 or more grouping variables.}
}
\value{
Returns a list of unique words by grouping variable.
}
\description{
Find unique words used by grouping variable.
}
\examples{
\dontrun{
dat <- pres_debates2012[pres_debates2012$time == "time 3", ]
with(dat, unique_by(dialogue, person))
with(pres_debates2012, unique_by(dialogue, list(time, person)))

with(DATA, unique_by(state, person))
}
}
\keyword{unique}

