% Generated by roxygen2 (4.0.0): do not edit by hand
\name{as.wfm.matrix}
\alias{as.wfm.DocumentTermMatrix}
\alias{as.wfm.TermDocumentMatrix}
\alias{as.wfm.data.frame}
\alias{as.wfm.default}
\alias{as.wfm.matrix}
\alias{as.wfm.wfdf}
\title{\code{as.wfm.matrix} - \code{matrix} method for \code{as.wfm} used to
convert matrices to a \code{wfm}.}
\usage{
\method{as.wfm}{matrix}(x)

\method{as.wfm}{default}(x)

\method{as.wfm}{TermDocumentMatrix}(x)

\method{as.wfm}{DocumentTermMatrix}(x)

\method{as.wfm}{data.frame}(x)

\method{as.wfm}{wfdf}(x)
}
\description{
\code{as.wfm.matrix} - \code{matrix} method for \code{as.wfm} used to
convert matrices to a \code{wfm}.

\code{as.wfm.default} - Default method for \code{as.wfm} used to
convert matrices to a \code{wfm}.

\code{as.wfm.TermDocumentMatrix} - \code{TermDocumentMatrix} method for
\code{as.wfm} used to a \code{TermDocumentMatrix} to a \code{wfm}.

\code{as.wfm.DocumentTermMatrix} - \code{DocumentTermMatrix} method for
\code{as.wfm} used to a \code{DocumentTermMatrix} to a \code{wfm}.

\code{as.wfm.data.frame} - data.frame method for \code{as.wfm} used to
convert matrices to a \code{wfm}.

\code{as.wfm.wfdf} - wfdf method for \code{as.wfm} used to
convert matrices to a \code{wfm}.
}

