% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Animate.R
\name{Animate}
\alias{Animate}
\title{Generic Animate Method}
\usage{
Animate(x, ...)
}
\arguments{
\item{x}{An animatable qdap object (e.g., \code{\link[qdap]{discourse_map}}).}

\item{\ldots}{Arguments passed to Animate method of other classes.}
}
\value{
Returns a plot object.
}
\description{
Animate select qdap objects.
}
\seealso{
\code{\link[qdap]{scores}},
\code{\link[qdap]{counts}},
\code{\link[qdap]{preprocessed}},
\code{\link[qdap]{proportions}}
}

