% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Remove Escaped Characters}
\usage{
clean(text.var)
}
\arguments{
\item{text.var}{The text variable}
}
\value{
Returns a vector of character strings with escaped characters removed.
}
\description{
Preprocess data to remove escaped characters
}
\examples{
\dontrun{
x <- "I go \\\\r
    to the \\\\tnext line"
x
clean(x)
}
}
\keyword{character}
\keyword{escaped}

