% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated_readability_index.R
\name{counts.flesch_kincaid}
\alias{counts.flesch_kincaid}
\title{Readability Measures}
\usage{
\method{counts}{flesch_kincaid}(x, ...)
}
\arguments{
\item{x}{The flesch_kincaid object.}

\item{\ldots}{ignored}
}
\description{
\code{counts.flesch_kincaid} - View counts from \code{\link[qdap]{flesch_kincaid}}.
}
\details{
flesch_kincaid Method for counts.
}
