% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos.R
\name{counts.pos}
\alias{counts.pos}
\title{Parts of Speech}
\usage{
\method{counts}{pos}(x, ...)
}
\arguments{
\item{x}{The \code{\link[qdap]{pos}} object.}

\item{\ldots}{ignored}
}
\description{
View pos counts.
}
\details{
pos Method for counts
}
