% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos.R
\name{counts.pos_by}
\alias{counts.pos_by}
\title{Parts of Speech}
\usage{
\method{counts}{pos_by}(x, ...)
}
\arguments{
\item{x}{The \code{\link[qdap]{pos_by}} object.}

\item{\ldots}{ignored}
}
\description{
View pos_by counts.
}
\details{
pos_by Method for counts
}
