% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subject_pronoun_type.R
\name{counts.subject_pronoun_type}
\alias{counts.subject_pronoun_type}
\title{Question Counts}
\usage{
\method{counts}{subject_pronoun_type}(x, ...)
}
\arguments{
\item{x}{The \code{\link[qdap]{subject_pronoun_type}} object.}

\item{\ldots}{ignored}
}
\description{
View subject_pronoun_type counts.
}
\details{
subject_pronoun_type Method for counts
}

