% Generated by roxygen2 (4.0.0): do not edit by hand
\name{df2tm_corpus}
\alias{df2tm_corpus}
\title{tm Package Compatibility Tools: Apply to or Convert to/from Term Document
Matrix or Document Term Matrix}
\usage{
df2tm_corpus(text.var, grouping.var = NULL, demographic.vars, ...)
}
\arguments{
\item{demographic.vars}{Additional demographic information about the grouping
variables.  This is a data.frame, list of equal length vectors, or a single
vector corresponding to the grouping variable/text variable.  This
information will be mapped to the DMetaData in the \code{\link[tm]{Corpus}}.}
}
\value{
\code{df2tm_corpus} - Converts a qdap oriented dataframe and returns
a \code{\link[tm]{Corpus}}.
}
\description{
\code{df2tm_corpus} - Convert a qdap dataframe to a tm package
\code{\link[tm]{Corpus}}.
}
\section{Warning}{
 The \code{df2tm_corpus} function is deprecated.  It will be
removed in a subsequent version of qdap.  Use \code{as.Corpus} instead.
}

