% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qdap-package.R
\docType{data}
\name{env.syl}
\alias{env.syl}
\title{Syllable Lookup Environment}
\format{A environment with the DICTIONARY data set.}
\usage{
data(env.syl)
}
\description{
A dataset containing a syllable lookup environment (see \code{DICTIONARY}).
}
\details{
For internal use.
}
\references{
\href{http://archive.ics.uci.edu/ml/machine-learning-databases/undocumented/connectionist-bench/nettalk/}{UCI Machine Learning Repository website}
}
\keyword{datasets}

