% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/name2sex.R
\name{name2sex}
\alias{name2sex}
\title{Names to Gender}
\usage{
name2sex(names.list, USE.NAMES = FALSE, ...)
}
\arguments{
\item{names.list}{Character vector containing first names.}

\item{USE.NAMES}{logical.  If \code{TRUE} names.list is used to name the
gender vector.}

\item{\ldots}{Other arguments passed to \code{\link[gender]{gender}}.}
}
\value{
Returns a vector of predicted gender (M/F) based on first name.
}
\description{
A wrapper for the \code{\link[gender]{gender}} function used to predict
gender based on first name.
}
\examples{
\dontrun{
name2sex(qcv(mary, jenn, linda, JAME, GABRIEL, OLIVA,
    tyler, jamie, JAMES, tyrone, cheryl, drew))
}
}
\seealso{
\code{\link[gender]{gender}}
}
\keyword{gender}
\keyword{name}

