% Generated by roxygen2 (4.0.0): do not edit by hand
\name{Network}
\alias{Network}
\alias{print.Network}
\title{Generic Network Method}
\usage{
Network(x, ...)

\method{print}{Network}(x, title = NA, title.color = "black",
  seed = sample(1:10000, 1), layout = layout.auto, legend = c(-0.5, -1.5,
  0.5, -1.45), legend.cex = 1, bg = NULL, legend.color = "black",
  vertex.color = "grey40", vertex.size = 20,
  vertex.label.color = "deepskyblue", vertex.label.cex = 1.1,
  edge.label.color = "black", edge.label.cex = 0.9, ...)
}
\arguments{
\item{x}{A select qdap object.}

\item{\ldots}{Arguments passed to Network method of other classes.}

\item{x}{The Network object.}

\item{title}{The title of the plot.  \code{NULL} eliminates title.  \code{NA}
uses title attribute of the Network object.}

\item{title.color}{The color of the title.}

\item{layout}{\pkg{igraph} \code{layout} to use.}

\item{seed}{The seed to use in plotting the graph.}

\item{legend}{The coordinates of the legend. See
\code{\link[plotrix]{color.legend}} for more information.}

\item{legend.cex}{character expansion factor. \code{NULL} and \code{NA} are
equivalent to 1.0. See \code{\link[graphics]{mtext}} for more information.}

\item{legend.color}{The text legend color for the network plot.}

\item{bg}{The color to be used for the background of the device region. See
\code{\link[graphics]{par}} for more information.}

\item{vertex.color}{The font family to be used for vertex labels.}

\item{vertex.size}{The size of the vertex.}

\item{vertex.label.color}{The color of the labels.}

\item{vertex.label.cex}{The font size for vertex labels.}

\item{edge.label.color}{The color for the edge labels.  Use \code{NA} to
remove.}

\item{edge.label.cex}{The font size of the edge labels.}

\item{\ldots}{Other Arguments passed to \code{\link[igraph]{plot.igraph}}.}
}
\value{
Returns a network plot.
}
\description{
Create a network plt for select qdap outputs.

Prints a Network object.
}
\note{
The output from \code{Network} is an \pkg{igraph} object and can be
altered and plotted directly using \pkg{igraph}.  The \pkg{qdap} \code{print}
method is offered as a quick approach to styling the figure.  For mor control
use \code{\link[igraph]{V}}, \code{\link[igraph]{E}}, and
\code{\link[igraph]{plot.igrapgh}}.
}

