% Generated by roxygen2 (4.0.0): do not edit by hand
\name{Network}
\alias{+.Network}
\alias{Network}
\title{Generic Network Method}
\usage{
Network(x, ...)

\method{+}{Network}(Network.obj, x)
}
\arguments{
\item{x}{A select qdap object.}

\item{\ldots}{Arguments passed to Network method of other classes.}

\item{Network.obj}{A object of class \code{Network}.}

\item{x}{A component to add to \code{Network.obj}}
}
\value{
Returns a network plot.
}
\description{
Create a network plt for select qdap outputs.

This operator allows you to add themes to a Network object.
}

