% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Network.R
\name{Network}
\alias{Network}
\title{Generic Network Method}
\usage{
Network(x, ...)
}
\arguments{
\item{x}{A select qdap object.}

\item{\ldots}{Arguments passed to Network method of other classes.}
}
\value{
Returns a network plot.
}
\description{
Create a network plot for select qdap outputs.
}

