% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discourse_map.R
\name{plot.animated_discourse_map}
\alias{plot.animated_discourse_map}
\title{Plots an animated_discourse_map  Object}
\usage{
\method{plot}{animated_discourse_map}(x, ...)
}
\arguments{
\item{x}{The animated_discourse_map  object.}

\item{\ldots}{Other arguments passed to \code{print.animated_discourse_map }.}
}
\description{
Plots an animated_discourse_map  object.
}

