% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formality.R
\name{plot.animated_formality}
\alias{plot.animated_formality}
\title{Plots a animated_formality  Object}
\usage{
\method{plot}{animated_formality}(x, ...)
}
\arguments{
\item{x}{The animated_formality  object.}

\item{\ldots}{Other arguments passed to \code{print.animated_formality }.}
}
\description{
Plots a animated_formality  object.
}
