% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lexical_classification.R
\name{plot.animated_lexical_classification}
\alias{plot.animated_lexical_classification}
\title{Plots an animated_lexical_classification  Object}
\usage{
\method{plot}{animated_lexical_classification}(x, ...)
}
\arguments{
\item{x}{The animated_lexical_classification  object.}

\item{\ldots}{Other arguments passed to \code{print.animated_lexical_classification }.}
}
\description{
Plots an animated_lexical_classification  object.
}

