% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.animated_polarity}
\alias{plot.animated_polarity}
\title{Plots a animated_polarity  Object}
\usage{
\method{plot}{animated_polarity}(x, ...)
}
\arguments{
\item{x}{The animated_polarity  object.}

\item{\ldots}{Other arguments passed to \code{print.animated_polarity }.}
}
\description{
Plots a animated_polarity  object.
}

