% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polarity.R
\name{plot.animated_polarity}
\alias{plot.animated_polarity}
\title{Plots an animated_polarity  Object}
\usage{
\method{plot}{animated_polarity}(x, ...)
}
\arguments{
\item{x}{The animated_polarity  object.}

\item{\ldots}{Other arguments passed to \code{print.animated_polarity }.}
}
\description{
Plots an animated_polarity  object.
}

