% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/automated_readability_index.R
\name{plot.automated_readability_index}
\alias{plot.automated_readability_index}
\title{Plots a automated_readability_index Object}
\usage{
\method{plot}{automated_readability_index}(x, ...)
}
\arguments{
\item{x}{The readability_score object.}

\item{\ldots}{ignored}
}
\description{
Plots a automated_readability_index object.
}

