% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/automated_readability_index.R
\name{plot.coleman_liau}
\alias{plot.coleman_liau}
\title{Plots a coleman_liau Object}
\usage{
\method{plot}{coleman_liau}(x, ...)
}
\arguments{
\item{x}{The readability_score object.}

\item{\ldots}{ignored}
}
\description{
Plots a coleman_liau object.
}

