% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/syllable_sum.R
\name{plot.combo_syllable_sum}
\alias{plot.combo_syllable_sum}
\title{Plots a combo_syllable_sum Object}
\usage{
\method{plot}{combo_syllable_sum}(x, ...)
}
\arguments{
\item{x}{The combo_syllable_sum object.}

\item{\ldots}{ignored}
}
\description{
Plots a combo_syllable_sum object.
}

