% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polarity.R
\name{plot.cumulative_animated_polarity}
\alias{plot.cumulative_animated_polarity}
\title{Plots a cumulative_animated_polarity Object}
\usage{
\method{plot}{cumulative_animated_polarity}(x, ...)
}
\arguments{
\item{x}{The cumulative_animated_polarity object.}

\item{\ldots}{ignored}
}
\description{
Plots a cumulative_animated_polarity object.
}

