% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/syllable_sum.R
\name{plot.cumulative_syllable_freq}
\alias{plot.cumulative_syllable_freq}
\title{Plots a cumulative_syllable_freq Object}
\usage{
\method{plot}{cumulative_syllable_freq}(x, ...)
}
\arguments{
\item{x}{The cumulative_syllable_freq object.}

\item{\ldots}{ignored}
}
\description{
Plots a cumulative_syllable_freq object.
}

