\name{plot.discourse_map}
\alias{plot.discourse_map}
\title{Plots a discourse_map Object}
\usage{
\method{plot}{discourse_map}(x, ...)
}
\arguments{
  \item{x}{The discourse_map object.}

  \item{\ldots}{Other arguments passed to
  \code{print.discourse_map}.}
}
\description{
Plots a discourse_map object.
}

