% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/freq_terms.R
\name{plot.freq_terms}
\alias{plot.freq_terms}
\title{Plots a freq_terms Object}
\usage{
\method{plot}{freq_terms}(x, plot = TRUE, ...)
}
\arguments{
\item{x}{The freq_terms object.}

\item{plot}{logical.  If \code{TRUE} the plot will automatically plot.
The user may wish to set to \code{FALSE} for use in knitr, sweave, etc.
to add additional plot layers.}

\item{\ldots}{ignored.}
}
\description{
Plots a freq_terms object.
}

