% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gantt.R
\name{plot.gantt}
\alias{plot.gantt}
\title{Plots a gantt object}
\usage{
\method{plot}{gantt}(x, base = FALSE, title = NULL, ...)
}
\arguments{
\item{x}{The sums_gantt object}

\item{base}{logical.  If \code{TRUE} prints in base graphics system.
If \code{FALSE} prints in ggplot graphics system.}

\item{title}{An optional title.}

\item{\ldots}{Other arguments passed to \code{gantt_wrap} or
\code{plot_gantt_base}}
}
\description{
Plots a gantt object.
}

