% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/automated_readability_index.R
\name{plot.linsear_write_count}
\alias{plot.linsear_write_count}
\title{Plots a linsear_write_count Object}
\usage{
\method{plot}{linsear_write_count}(x, ...)
}
\arguments{
\item{x}{The linsear_write_count object.}

\item{\ldots}{ignored}
}
\description{
Plots a linsear_write_count object.
}

