% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/syllable_sum.R
\name{plot.syllable_freq}
\alias{plot.syllable_freq}
\title{Plots a syllable_freq Object}
\usage{
\method{plot}{syllable_freq}(x, ...)
}
\arguments{
\item{x}{The syllable_freq object.}

\item{\ldots}{ignored}
}
\description{
Plots a syllable_freq object.
}

