% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plot.table_score}
\alias{plot.table_score}
\title{Plots a table_score Object}
\usage{
\method{plot}{table_score}(x, values = TRUE, high = "red", ...)
}
\arguments{
\item{x}{The table_score object.}

\item{high}{The color to be used for higher values.}

\item{values}{logical.  If \code{TRUE} the cell values will be included on
the heatmap.}

\item{\ldots}{Other arguments passed to \code{\link[qdap]{qheat}}.}
}
\description{
Plots a table_score object.
}

