% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/word_stats.R
\name{plot.word_stats_counts}
\alias{plot.word_stats_counts}
\title{Plots a word_stats_counts Object}
\usage{
\method{plot}{word_stats_counts}(x, alpha = 0.3, ...)
}
\arguments{
\item{x}{The word_stats_counts object.}

\item{alpha}{The alpha level to use for points.}

\item{\ldots}{ignored}
}
\description{
Plots a word_stats_counts object.
}

