% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formality.R
\name{preprocessed.formality}
\alias{preprocessed.formality}
\title{Formality}
\usage{
\method{preprocessed}{formality}(x, ...)
}
\arguments{
\item{x}{The \code{\link[qdap]{formality}} object.}

\item{\ldots}{ignored}
}
\description{
View formality preprocessed.
}
\details{
formality Method for preprocessed
}
