% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lexical_classification.R
\name{preprocessed.lexical_classification}
\alias{preprocessed.lexical_classification}
\title{Lexical Classification}
\usage{
\method{preprocessed}{lexical_classification}(x, ...)
}
\arguments{
\item{x}{The lexical_classification object.}

\item{\ldots}{ignored}
}
\description{
\code{preprocessed.lexical_classification} - View preprocessed from \code{\link[qdap]{lexical_classification}}.
}
\details{
lexical_classification Method for preprocessed.
}

