% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos.R
\name{preprocessed.pos}
\alias{preprocessed.pos}
\title{Parts of Speech}
\usage{
\method{preprocessed}{pos}(x, ...)
}
\arguments{
\item{x}{The \code{\link[qdap]{pos}} object.}

\item{\ldots}{ignored}
}
\description{
View pos preprocessed.
}
\details{
pos Method for preprocessed
}
