% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pronoun_type.R
\name{preprocessed.pronoun_type}
\alias{preprocessed.pronoun_type}
\title{Question Counts}
\usage{
\method{preprocessed}{pronoun_type}(x, ...)
}
\arguments{
\item{x}{The pronoun_type object.}

\item{\ldots}{ignored}
}
\description{
View \code{\link[qdap]{pronoun_type}} preprocessed.
}
\details{
pronoun_type Method for preprocessed
}

