% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subject_pronoun_type.R
\name{preprocessed.subject_pronoun_type}
\alias{preprocessed.subject_pronoun_type}
\title{Question Counts}
\usage{
\method{preprocessed}{subject_pronoun_type}(x, ...)
}
\arguments{
\item{x}{The subject_pronoun_type object.}

\item{\ldots}{ignored}
}
\description{
View \code{\link[qdap]{subject_pronoun_type}} preprocessed.
}
\details{
subject_pronoun_type Method for preprocessed
}

