% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/word_position.R
\name{preprocessed.word_position}
\alias{preprocessed.word_position}
\title{Word Position}
\usage{
\method{preprocessed}{word_position}(x, ...)
}
\arguments{
\item{x}{The word_position object.}

\item{\ldots}{ignored}
}
\description{
View \code{\link[qdap]{word_position}} preprocessed.
}
\details{
word_position Method for preprocessed
}

