% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Animate.character.R
\name{print.animated_character}
\alias{print.animated_character}
\title{Prints an animated_character  Object}
\usage{
\method{print}{animated_character}(x, pause = 0, ...)
}
\arguments{
\item{x}{The animated_character  object.}

\item{pause}{The length of time to pause between plots.}

\item{\ldots}{ignored.}
}
\description{
Prints an animated_character  object.
}

