% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Search.R
\name{print.boolean_qdap}
\alias{print.boolean_qdap}
\title{Prints a boolean_qdap object}
\usage{
\method{print}{boolean_qdap}(x, ...)
}
\arguments{
\item{x}{The boolean_qdap object}

\item{\ldots}{ignored}
}
\description{
Prints a boolean_qdap object
}

