% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colsplit2df.R
\name{print.colsplit2df}
\alias{print.colsplit2df}
\title{Prints a colsplit2df Object.}
\usage{
\method{print}{colsplit2df}(x, ...)
}
\arguments{
\item{x}{The colsplit2df object}

\item{\ldots}{ignored}
}
\description{
Prints a colsplit2df object.
}

