% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/syllable_sum.R
\name{print.cumulative_combo_syllable_sum}
\alias{print.cumulative_combo_syllable_sum}
\title{Prints a cumulative_combo_syllable_sum Object}
\usage{
\method{print}{cumulative_combo_syllable_sum}(x, ...)
}
\arguments{
\item{x}{The cumulative_combo_syllable_sum object.}

\item{\ldots}{ignored}
}
\description{
Prints a cumulative_combo_syllable_sum object.
}

