% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/end_mark.R
\name{print.cumulative_end_mark}
\alias{print.cumulative_end_mark}
\title{Prints a cumulative_end_mark Object}
\usage{
\method{print}{cumulative_end_mark}(x, ...)
}
\arguments{
\item{x}{The cumulative_end_mark object.}

\item{\ldots}{ignored}
}
\description{
Prints a cumulative_end_mark  object.
}

