% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/syllable_sum.R
\name{print.cumulative_syllable_freq}
\alias{print.cumulative_syllable_freq}
\title{Prints a cumulative_syllable_freqObject}
\usage{
\method{print}{cumulative_syllable_freq}(x, ...)
}
\arguments{
\item{x}{The cumulative_syllable_freqobject.}

\item{\ldots}{ignored}
}
\description{
Prints a cumulative_syllable_freq object.
}

