% Generated by roxygen2 (4.0.0): do not edit by hand
\name{print.discourse_map}
\alias{print.discourse_map}
\title{Prints a discourse_map Object}
\usage{
\method{print}{discourse_map}(x, edge.curved = TRUE, title = NULL, ...)
}
\arguments{
\item{x}{The discourse_map object.}

\item{edge.curved}{logical.  If \code{TRUE} edges are plotted with curves.}

\item{title}{The title of the plot.}

\item{\ldots}{Other Arguments passed to \code{\link[igraph]{plot.igraph}}.}
}
\description{
Prints a discourse_map object.
}

