\name{print.dissimilarity}
\alias{print.dissimilarity}
\title{Prints a dissimilarity object}
\usage{
  \method{print}{dissimilarity} (x, digits = 3, ...)
}
\arguments{
  \item{x}{The dissimilarity object}

  \item{digits}{Number of decimal places to print.}

  \item{\ldots}{ignored}
}
\description{
  Prints a dissimilarity object.
}

