% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dissimilarity.R
\name{print.Dissimilarity}
\alias{print.Dissimilarity}
\title{Prints a Dissimilarity object}
\usage{
\method{print}{Dissimilarity}(x, digits = 3, ...)
}
\arguments{
\item{x}{The Dissimilarity object}

\item{digits}{Number of decimal places to print.}

\item{\ldots}{ignored}
}
\description{
Prints a Dissimilarity object.
}

