\name{print.diversity}
\alias{print.diversity}
\title{Prints a diversity object}
\usage{
  \method{print}{diversity} (x, digits = 3, ...)
}
\arguments{
  \item{x}{The diversity object}

  \item{digits}{Number of decimal places to print.}

  \item{\ldots}{ignored}
}
\description{
  Prints a diversity object.
}

