% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversity.R
\name{print.diversity}
\alias{print.diversity}
\title{Prints a diversity object}
\usage{
\method{print}{diversity}(x, digits = 3, ...)
}
\arguments{
\item{x}{The diversity object}

\item{digits}{Number of decimal places to print.}

\item{\ldots}{ignored}
}
\description{
Prints a diversity object.
}

