% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/end_mark.R
\name{print.end_mark_by}
\alias{print.end_mark_by}
\title{Prints an end_mark_by object}
\usage{
\method{print}{end_mark_by}(x, ...)
}
\arguments{
\item{x}{The end_mark_by object}

\item{\ldots}{ignored}
}
\description{
Prints an end_mark_by object
}

