% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phrase_net.R
\name{print.phrase_net}
\alias{print.phrase_net}
\title{Prints a phrase_net Object}
\usage{
\method{print}{phrase_net}(x, edge.curved = TRUE, ...)
}
\arguments{
\item{x}{The phrase_net object.}

\item{edge.curved}{logical.  If \code{TRUE} edges are plotted with curves.}

\item{\ldots}{Other Arguments passed to \code{\link[igraph]{plot.igraph}}.}
}
\description{
Prints a phrase_net object.
}

