\name{print.polarity}
\alias{print.polarity}
\title{Prints a polarity Object}
\usage{
  \method{print}{polarity} (x, digits = NULL, ...)
}
\arguments{
  \item{x}{The polarity object.}

  \item{digits}{Number of decimal places to print.}

  \item{\ldots}{ignored}
}
\description{
  Prints a polarity object.
}

