% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polarity.R
\name{print.polarity}
\alias{print.polarity}
\title{Prints an polarity Object}
\usage{
\method{print}{polarity}(x, digits = 3, ...)
}
\arguments{
\item{x}{The polarity object.}

\item{digits}{The number of digits displayed if \code{values} is \code{TRUE}.}

\item{\ldots}{ignored}
}
\description{
Prints an polarity object.
}

