% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/syllable_sum.R
\name{print.polysyllable_sum}
\alias{print.polysyllable_sum}
\title{Prints an polysyllable_sum object}
\usage{
\method{print}{polysyllable_sum}(x, ...)
}
\arguments{
\item{x}{The polysyllable_sum object}

\item{\ldots}{ignored}
}
\description{
Prints an polysyllable_sum object
}

