% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/table_methods.R
\name{print.table_proportion}
\alias{print.table_proportion}
\title{Prints a table_proportion object}
\usage{
\method{print}{table_proportion}(x, ...)
}
\arguments{
\item{x}{The table_proportion object}

\item{\ldots}{ignored}
}
\description{
Prints a table_proportion object
}

