% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/table_methods.R
\name{print.table_score}
\alias{print.table_score}
\title{Prints a table_score object}
\usage{
\method{print}{table_score}(x, ...)
}
\arguments{
\item{x}{The table_score object}

\item{\ldots}{ignored}
}
\description{
Prints a table_score object
}

