% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wfm.R
\name{print.wfm}
\alias{print.wfm}
\title{Prints a wfm Object}
\usage{
\method{print}{wfm}(x, digits = 3, width = 10000, ...)
}
\arguments{
\item{x}{The wfm object.}

\item{digits}{The number of digits displayed if \code{values} is \code{TRUE}.}

\item{width}{The width to temporarily set for printing (default = 10000).
See \code{\link[base]{options}} for more.}

\item{\ldots}{ignored}
}
\description{
Prints a wfm object.
}

