% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/word_associate.R
\name{print.word_associate}
\alias{print.word_associate}
\title{Prints a word_associate object}
\usage{
\method{print}{word_associate}(x, ...)
}
\arguments{
\item{x}{The word_associate object}

\item{\ldots}{ignored}
}
\description{
Prints a word_associate object.
}

