% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/word_position.R
\name{print.word_position}
\alias{print.word_position}
\title{Prints a word_position object.}
\usage{
\method{print}{word_position}(x, ...)
}
\arguments{
\item{x}{The word_position object}

\item{\ldots}{Values passed to \code{plot.word_position}}
}
\description{
Prints a word_position object.
}

