% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/word_proximity.R
\name{print.word_proximity}
\alias{print.word_proximity}
\title{Prints a word_proximity object}
\usage{
\method{print}{word_proximity}(x, digits = NULL, ...)
}
\arguments{
\item{x}{The word_proximity object}

\item{digits}{The number of digits to print}

\item{\ldots}{ignored}
}
\description{
Prints a word_proximity object
}

