% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/word_stats.R
\name{print.word_stats_counts}
\alias{print.word_stats_counts}
\title{Prints a word_stats_counts object}
\usage{
\method{print}{word_stats_counts}(x, ...)
}
\arguments{
\item{x}{The word_stats_counts object}

\item{\ldots}{ignored}
}
\description{
Prints a word_stats_counts object
}

