% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/object_pronoun_type.R
\name{proportions.object_pronoun_type}
\alias{proportions.object_pronoun_type}
\title{Question Counts}
\usage{
\method{proportions}{object_pronoun_type}(x, ...)
}
\arguments{
\item{x}{The object_pronoun_type object.}

\item{\ldots}{ignored}
}
\description{
View \code{\link[qdap]{object_pronoun_type}} proportions.
}
\details{
object_pronoun_type Method for proportions
}

