% Generated by roxygen2 (4.0.1): do not edit by hand
\name{qdap_df}
\alias{Text}
\alias{Text<-}
\alias{qdap_df}
\title{Create qdap Specific Data Structure}
\usage{
qdap_df(dataframe, text.var)

Text(object)

Text(object) <- value
}
\arguments{
\item{dataframe}{A \code{\link[base]{data.frame}} with a text variable.
Generally, \code{\link[qdap]{sentSplit}} should be run first
(\code{\link[qdap]{sentSplit}}) actually produces a
\code{\link[base]{data.frame}} that is of the class \code{"qdap_df"}).}

\item{text.var}{The name of the \code{text.var} column.}

\item{object}{A \code{\link[base]{data.frame}} of the class \code{"qdap_df"}.}

\item{value}{A character string of the updated \code{text.var} column.}
}
\value{
Returns a \code{\link[base]{data.frame}} of the class \code{"qdap_df"}.
}
\description{
Creating this \pkg{qdap} specific data structure enables short hand with
subsequent \pkg{qdap} function calls that utilize the \code{text.var}
argument.  Combined with the \code{\%&\%} operator, the user need not specifiy
a data set or the \code{text.var} argument.

Change text.var column of a qdap_df object.
}
\examples{
\dontrun{
dat <- qdap_df(DATA, state)
dat \%&\% trans_cloud(grouping.var=person)
dat \%&\% trans_cloud(grouping.var=person, text.var=stemmer(DATA$state))
dat \%&\% termco(grouping.var=person, match.list=list("fun", "computer"))
class(dat)

dat2 <- sentSplit(DATA, "state", stem.col = TRUE)
class(dat2)
dat2 \%&\% trans_cloud()
Text(dat2)
## change the `text.var` column
Text(dat2) <- "stem.text"
dat2 \%&\% trans_cloud()
}
}
\references{
Inspired by \pkg{dplyr}'s \code{\link[dplyr]{tbl_df}} structure.
}
\seealso{
\code{\link[qdap]{sentSplit}}
}
\keyword{Text}
\keyword{data}
\keyword{structure}

