% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qdap-package.R
\docType{data}
\name{raj.act.4}
\alias{raj.act.4}
\title{Romeo and Juliet: Act 4}
\format{A data frame with 115 rows and 2 variables}
\usage{
data(raj.act.4)
}
\description{
A dataset containing Romeo and Juliet: Act 4.
}
\details{
\itemize{
  \item person. Character in the play
  \item dialogue. The spoken dialogue
}
}
\references{
\url{http://shakespeare.mit.edu/romeo_juliet/full.html}
}
\keyword{datasets}

