% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/word_count.R
\name{scores.character_table}
\alias{scores.character_table}
\title{Term Counts}
\usage{
\method{scores}{character_table}(x, ...)
}
\arguments{
\item{x}{The \code{\link[qdap]{character_table}} object.}

\item{\ldots}{ignored}
}
\description{
View character_table scores.
}
\details{
character_table Method for scores
}

