% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated_readability_index.R
\name{scores.flesch_kincaid}
\alias{scores.flesch_kincaid}
\title{Readability Measures}
\usage{
\method{scores}{flesch_kincaid}(x, ...)
}
\arguments{
\item{x}{The flesch_kincaid object.}

\item{\ldots}{ignored}
}
\description{
\code{scores.flesch_kincaid} - View scores from \code{\link[qdap]{flesch_kincaid}}.
}
\details{
flesch_kincaid Method for scores
}
