% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formality.R
\name{scores.formality}
\alias{scores.formality}
\title{Formality}
\usage{
\method{scores}{formality}(x, ...)
}
\arguments{
\item{x}{The \code{\link[qdap]{formality}} object.}

\item{\ldots}{ignored}
}
\description{
View formality scores.
}
\details{
formality Method for scores
}
