% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated_readability_index.R
\name{scores.fry}
\alias{scores.fry}
\title{Readability Measures}
\usage{
\method{scores}{fry}(x, ...)
}
\arguments{
\item{x}{The fry object.}

\item{\ldots}{ignored}
}
\description{
\code{scores.fry} - View scores from \code{\link[qdap]{fry}}.
}
\details{
fry Method for scores
}
