% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lexical_classification.R
\name{scores.lexical_classification}
\alias{scores.lexical_classification}
\title{Lexical Classification}
\usage{
\method{scores}{lexical_classification}(x, ...)
}
\arguments{
\item{x}{The lexical_classification object.}

\item{\ldots}{ignored}
}
\description{
\code{scores.lexical_classification} - View scores from \code{\link[qdap]{lexical_classification}}.
}
\details{
lexical_classification Method for scores
}

