% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated_readability_index.R
\name{scores.linsear_write}
\alias{scores.linsear_write}
\title{Readability Measures}
\usage{
\method{scores}{linsear_write}(x, ...)
}
\arguments{
\item{x}{The linsear_write object.}

\item{\ldots}{ignored}
}
\description{
\code{scores.linsear_write} - View scores from \code{\link[qdap]{linsear_write}}.
}
\details{
linsear_write Method for scores
}
