% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scores.question_type}
\alias{scores.question_type}
\title{Question Counts}
\usage{
\method{scores}{question_type}(x, ...)
}
\arguments{
\item{x}{The \code{\link[qdap]{question_type}} object.}

\item{\ldots}{ignored}
}
\description{
View question_type scores.
}
\details{
question_type Method for scores
}

