% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spaste.R
\name{spaste}
\alias{spaste}
\title{Add Leading/Trailing Spaces}
\usage{
spaste(terms, trailing = TRUE, leading = TRUE)
}
\arguments{
\item{terms}{A character vector of terms to insert trailing and/or leading
spaces.}

\item{trailing}{logical.  If \code{TRUE} inserts a trailing space in the
terms.}

\item{leading}{logical.  If \code{TRUE} inserts a leading space in the terms.}
}
\value{
Returns a character vector with trailing and/or leading spaces.
}
\description{
Adds trailing and/or leading spaces to a vector of terms.
}
\examples{
\dontrun{
spaste(Top25Words)
spaste(Top25Words, FALSE)
spaste(Top25Words, trailing = TRUE, leading = FALSE) #or
spaste(Top25Words, , FALSE)
}
}

