% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.tdm.R
\name{t.DocumentTermMatrix}
\alias{t.DocumentTermMatrix}
\title{Transposes a DocumentTermMatrix object}
\usage{
\method{t}{DocumentTermMatrix}(x, ...)
}
\arguments{
\item{x}{The DocumentTermMatrix object}

\item{\ldots}{ignored}
}
\description{
Transposes a DocumentTermMatrix object
}

